Restructure app components.
