+++ jhub_apps/static/js/index.js
@@ -78,8 +78,8 @@ Error generating stack: `+i.message+`
 	Copyright (c) 2018 Jed Watson.
 	Licensed under the MIT License (MIT), see
 	http://jedwatson.github.io/classnames
-*/(function(e){(function(){var t={}.hasOwnProperty;function n(){for(var r=[],o=0;o<arguments.length;o++){var i=arguments[o];if(i){var s=typeof i;if(s==="string"||s==="number")r.push(i);else if(Array.isArray(i)){if(i.length){var l=n.apply(null,i);l&&r.push(l)}}else if(s==="object"){if(i.toString!==Object.prototype.toString&&!i.toString.toString().includes("[native code]")){r.push(i.toString());continue}for(var a in i)t.call(i,a)&&i[a]&&r.push(a)}}}return r.join(" ")}e.exports?(n.default=n,e.exports=n):window.classNames=n})()})(pg);var Qk=pg.exports;const Yn=mc(Qk),qk=({id:e,type:t="info",children:n})=>{const r=Yn("alert",{"alert-success":t==="success","alert-warning":t==="warning","alert-error":t==="error","alert-info":t==="info"});return C.jsx("div",{id:e,className:r,children:C.jsx("div",{className:"alert-body",children:C.jsx("p",{className:"alert-text",children:n})})})};function mg(e,t){return function(){return e.apply(t,arguments)}}const{toString:Kk}=Object.prototype,{getPrototypeOf:Df}=Object,ta=(e=>t=>{const n=Kk.call(t);return e[n]||(e[n]=n.slice(8,-1).toLowerCase())})(Object.create(null)),sn=e=>(e=e.toLowerCase(),t=>ta(t)===e),na=e=>t=>typeof t===e,{isArray:po}=Array,gi=na("undefined");function Gk(e){return e!==null&&!gi(e)&&e.constructor!==null&&!gi(e.constructor)&&Dt(e.constructor.isBuffer)&&e.constructor.isBuffer(e)}const yg=sn("ArrayBuffer");function Yk(e){let t;return typeof ArrayBuffer<"u"&&ArrayBuffer.isView?t=ArrayBuffer.isView(e):t=e&&e.buffer&&yg(e.buffer),t}const Jk=na("string"),Dt=na("function"),vg=na("number"),ra=e=>e!==null&&typeof e=="object",Xk=e=>e===!0||e===!1,Es=e=>{if(ta(e)!=="object")return!1;const t=Df(e);return(t===null||t===Object.prototype||Object.getPrototypeOf(t)===null)&&!(Symbol.toStringTag in e)&&!(Symbol.iterator in e)},Zk=sn("Date"),eN=sn("File"),tN=sn("Blob"),nN=sn("FileList"),rN=e=>ra(e)&&Dt(e.pipe),oN=e=>{let t;return e&&(typeof FormData=="function"&&e instanceof FormData||Dt(e.append)&&((t=ta(e))==="formdata"||t==="object"&&Dt(e.toString)&&e.toString()==="[object FormData]"))},iN=sn("URLSearchParams"),sN=e=>e.trim?e.trim():e.replace(/^[\s\uFEFF\xA0]+|[\s\uFEFF\xA0]+$/g,"");function Oi(e,t,{allOwnKeys:n=!1}={}){if(e===null||typeof e>"u")return;let r,o;if(typeof e!="object"&&(e=[e]),po(e))for(r=0,o=e.length;r<o;r++)t.call(null,e[r],r,e);else{const i=n?Object.getOwnPropertyNames(e):Object.keys(e),s=i.length;let l;for(r=0;r<s;r++)l=i[r],t.call(null,e[l],l,e)}}function gg(e,t){t=t.toLowerCase();const n=Object.keys(e);let r=n.length,o;for(;r-- >0;)if(o=n[r],t===o.toLowerCase())return o;return null}const Sg=(()=>typeof globalThis<"u"?globalThis:typeof self<"u"?self:typeof window<"u"?window:global)(),wg=e=>!gi(e)&&e!==Sg;function ac(){const{caseless:e}=wg(this)&&this||{},t={},n=(r,o)=>{const i=e&&gg(t,o)||o;Es(t[i])&&Es(r)?t[i]=ac(t[i],r):Es(r)?t[i]=ac({},r):po(r)?t[i]=r.slice():t[i]=r};for(let r=0,o=arguments.length;r<o;r++)arguments[r]&&Oi(arguments[r],n);return t}const lN=(e,t,n,{allOwnKeys:r}={})=>(Oi(t,(o,i)=>{n&&Dt(o)?e[i]=mg(o,n):e[i]=o},{allOwnKeys:r}),e),aN=e=>(e.charCodeAt(0)===65279&&(e=e.slice(1)),e),uN=(e,t,n,r)=>{e.prototype=Object.create(t.prototype,r),e.prototype.constructor=e,Object.defineProperty(e,"super",{value:t.prototype}),n&&Object.assign(e.prototype,n)},cN=(e,t,n,r)=>{let o,i,s;const l={};if(t=t||{},e==null)return t;do{for(o=Object.getOwnPropertyNames(e),i=o.length;i-- >0;)s=o[i],(!r||r(s,e,t))&&!l[s]&&(t[s]=e[s],l[s]=!0);e=n!==!1&&Df(e)}while(e&&(!n||n(e,t))&&e!==Object.prototype);return t},fN=(e,t,n)=>{e=String(e),(n===void 0||n>e.length)&&(n=e.length),n-=t.length;const r=e.indexOf(t,n);return r!==-1&&r===n},dN=e=>{if(!e)return null;if(po(e))return e;let t=e.length;if(!vg(t))return null;const n=new Array(t);for(;t-- >0;)n[t]=e[t];return n},hN=(e=>t=>e&&t instanceof e)(typeof Uint8Array<"u"&&Df(Uint8Array)),pN=(e,t)=>{const r=(e&&e[Symbol.iterator]).call(e);let o;for(;(o=r.next())&&!o.done;){const i=o.value;t.call(e,i[0],i[1])}},mN=(e,t)=>{let n;const r=[];for(;(n=e.exec(t))!==null;)r.push(n);return r},yN=sn("HTMLFormElement"),vN=e=>e.toLowerCase().replace(/[-_\s]([a-z\d])(\w*)/g,function(n,r,o){return r.toUpperCase()+o}),Wh=(({hasOwnProperty:e})=>(t,n)=>e.call(t,n))(Object.prototype),gN=sn("RegExp"),_g=(e,t)=>{const n=Object.getOwnPropertyDescriptors(e),r={};Oi(n,(o,i)=>{let s;(s=t(o,i,e))!==!1&&(r[i]=s||o)}),Object.defineProperties(e,r)},SN=e=>{_g(e,(t,n)=>{if(Dt(e)&&["arguments","caller","callee"].indexOf(n)!==-1)return!1;const r=e[n];if(Dt(r)){if(t.enumerable=!1,"writable"in t){t.writable=!1;return}t.set||(t.set=()=>{throw Error("Can not rewrite read-only method '"+n+"'")})}})},wN=(e,t)=>{const n={},r=o=>{o.forEach(i=>{n[i]=!0})};return po(e)?r(e):r(String(e).split(t)),n},_N=()=>{},RN=(e,t)=>(e=+e,Number.isFinite(e)?e:t),za="abcdefghijklmnopqrstuvwxyz",Qh="0123456789",Rg={DIGIT:Qh,ALPHA:za,ALPHA_DIGIT:za+za.toUpperCase()+Qh},EN=(e=16,t=Rg.ALPHA_DIGIT)=>{let n="";const{length:r}=t;for(;e--;)n+=t[Math.random()*r|0];return n};function CN(e){return!!(e&&Dt(e.append)&&e[Symbol.toStringTag]==="FormData"&&e[Symbol.iterator])}const xN=e=>{const t=new Array(10),n=(r,o)=>{if(ra(r)){if(t.indexOf(r)>=0)return;if(!("toJSON"in r)){t[o]=r;const i=po(r)?[]:{};return Oi(r,(s,l)=>{const a=n(s,o+1);!gi(a)&&(i[l]=a)}),t[o]=void 0,i}}return r};return n(e,0)},TN=sn("AsyncFunction"),kN=e=>e&&(ra(e)||Dt(e))&&Dt(e.then)&&Dt(e.catch),P={isArray:po,isArrayBuffer:yg,isBuffer:Gk,isFormData:oN,isArrayBufferView:Yk,isString:Jk,isNumber:vg,isBoolean:Xk,isObject:ra,isPlainObject:Es,isUndefined:gi,isDate:Zk,isFile:eN,isBlob:tN,isRegExp:gN,isFunction:Dt,isStream:rN,isURLSearchParams:iN,isTypedArray:hN,isFileList:nN,forEach:Oi,merge:ac,extend:lN,trim:sN,stripBOM:aN,inherits:uN,toFlatObject:cN,kindOf:ta,kindOfTest:sn,endsWith:fN,toArray:dN,forEachEntry:pN,matchAll:mN,isHTMLForm:yN,hasOwnProperty:Wh,hasOwnProp:Wh,reduceDescriptors:_g,freezeMethods:SN,toObjectSet:wN,toCamelCase:vN,noop:_N,toFiniteNumber:RN,findKey:gg,global:Sg,isContextDefined:wg,ALPHABET:Rg,generateString:EN,isSpecCompliantForm:CN,toJSONObject:xN,isAsyncFn:TN,isThenable:kN};function me(e,t,n,r,o){Error.call(this),Error.captureStackTrace?Error.captureStackTrace(this,this.constructor):this.stack=new Error().stack,this.message=e,this.name="AxiosError",t&&(this.code=t),n&&(this.config=n),r&&(this.request=r),o&&(this.response=o)}P.inherits(me,Error,{toJSON:function(){return{message:this.message,name:this.name,description:this.description,number:this.number,fileName:this.fileName,lineNumber:this.lineNumber,columnNumber:this.columnNumber,stack:this.stack,config:P.toJSONObject(this.config),code:this.code,status:this.response&&this.response.status?this.response.status:null}}});const Eg=me.prototype,Cg={};["ERR_BAD_OPTION_VALUE","ERR_BAD_OPTION","ECONNABORTED","ETIMEDOUT","ERR_NETWORK","ERR_FR_TOO_MANY_REDIRECTS","ERR_DEPRECATED","ERR_BAD_RESPONSE","ERR_BAD_REQUEST","ERR_CANCELED","ERR_NOT_SUPPORT","ERR_INVALID_URL"].forEach(e=>{Cg[e]={value:e}});Object.defineProperties(me,Cg);Object.defineProperty(Eg,"isAxiosError",{value:!0});me.from=(e,t,n,r,o,i)=>{const s=Object.create(Eg);return P.toFlatObject(e,s,function(a){return a!==Error.prototype},l=>l!=="isAxiosError"),me.call(s,e.message,t,n,r,o),s.cause=e,s.name=e.name,i&&Object.assign(s,i),s};const NN=null;function uc(e){return P.isPlainObject(e)||P.isArray(e)}function xg(e){return P.endsWith(e,"[]")?e.slice(0,-2):e}function qh(e,t,n){return e?e.concat(t).map(function(o,i){return o=xg(o),!n&&i?"["+o+"]":o}).join(n?".":""):t}function AN(e){return P.isArray(e)&&!e.some(uc)}const LN=P.toFlatObject(P,{},null,function(t){return/^is[A-Z]/.test(t)});function oa(e,t,n){if(!P.isObject(e))throw new TypeError("target must be an object");t=t||new FormData,n=P.toFlatObject(n,{metaTokens:!0,dots:!1,indexes:!1},!1,function(v,T){return!P.isUndefined(T[v])});const r=n.metaTokens,o=n.visitor||f,i=n.dots,s=n.indexes,a=(n.Blob||typeof Blob<"u"&&Blob)&&P.isSpecCompliantForm(t);if(!P.isFunction(o))throw new TypeError("visitor must be a function");function u(p){if(p===null)return"";if(P.isDate(p))return p.toISOString();if(!a&&P.isBlob(p))throw new me("Blob is not supported. Use a Buffer instead.");return P.isArrayBuffer(p)||P.isTypedArray(p)?a&&typeof Blob=="function"?new Blob([p]):Buffer.from(p):p}function f(p,v,T){let m=p;if(p&&!T&&typeof p=="object"){if(P.endsWith(v,"{}"))v=r?v:v.slice(0,-2),p=JSON.stringify(p);else if(P.isArray(p)&&AN(p)||(P.isFileList(p)||P.endsWith(v,"[]"))&&(m=P.toArray(p)))return v=xg(v),m.forEach(function(y,_){!(P.isUndefined(y)||y===null)&&t.append(s===!0?qh([v],_,i):s===null?v:v+"[]",u(y))}),!1}return uc(p)?!0:(t.append(qh(T,v,i),u(p)),!1)}const c=[],h=Object.assign(LN,{defaultVisitor:f,convertValue:u,isVisitable:uc});function S(p,v){if(!P.isUndefined(p)){if(c.indexOf(p)!==-1)throw Error("Circular reference detected in "+v.join("."));c.push(p),P.forEach(p,function(m,d){(!(P.isUndefined(m)||m===null)&&o.call(t,m,P.isString(d)?d.trim():d,v,h))===!0&&S(m,v?v.concat(d):[d])}),c.pop()}}if(!P.isObject(e))throw new TypeError("data must be an object");return S(e),t}function Kh(e){const t={"!":"%21","'":"%27","(":"%28",")":"%29","~":"%7E","%20":"+","%00":"\0"};return encodeURIComponent(e).replace(/[!'()~]|%20|%00/g,function(r){return t[r]})}function Mf(e,t){this._pairs=[],e&&oa(e,this,t)}const Tg=Mf.prototype;Tg.append=function(t,n){this._pairs.push([t,n])};Tg.toString=function(t){const n=t?function(r){return t.call(this,r,Kh)}:Kh;return this._pairs.map(function(o){return n(o[0])+"="+n(o[1])},"").join("&")};function PN(e){return encodeURIComponent(e).replace(/%3A/gi,":").replace(/%24/g,"$").replace(/%2C/gi,",").replace(/%20/g,"+").replace(/%5B/gi,"[").replace(/%5D/gi,"]")}function kg(e,t,n){if(!t)return e;const r=n&&n.encode||PN,o=n&&n.serialize;let i;if(o?i=o(t,n):i=P.isURLSearchParams(t)?t.toString():new Mf(t,n).toString(r),i){const s=e.indexOf("#");s!==-1&&(e=e.slice(0,s)),e+=(e.indexOf("?")===-1?"?":"&")+i}return e}class ON{constructor(){this.handlers=[]}use(t,n,r){return this.handlers.push({fulfilled:t,rejected:n,synchronous:r?r.synchronous:!1,runWhen:r?r.runWhen:null}),this.handlers.length-1}eject(t){this.handlers[t]&&(this.handlers[t]=null)}clear(){this.handlers&&(this.handlers=[])}forEach(t){P.forEach(this.handlers,function(r){r!==null&&t(r)})}}const Gh=ON,Ng={silentJSONParsing:!0,forcedJSONParsing:!0,clarifyTimeoutError:!1},bN=typeof URLSearchParams<"u"?URLSearchParams:Mf,FN=typeof FormData<"u"?FormData:null,DN=typeof Blob<"u"?Blob:null,MN={isBrowser:!0,classes:{URLSearchParams:bN,FormData:FN,Blob:DN},protocols:["http","https","file","blob","url","data"]},Ag=typeof window<"u"&&typeof document<"u",UN=(e=>Ag&&["ReactNative","NativeScript","NS"].indexOf(e)<0)(typeof navigator<"u"&&navigator.product),IN=(()=>typeof WorkerGlobalScope<"u"&&self instanceof WorkerGlobalScope&&typeof self.importScripts=="function")(),VN=Object.freeze(Object.defineProperty({__proto__:null,hasBrowserEnv:Ag,hasStandardBrowserEnv:UN,hasStandardBrowserWebWorkerEnv:IN},Symbol.toStringTag,{value:"Module"})),Zt={...VN,...MN};function $N(e,t){return oa(e,new Zt.classes.URLSearchParams,Object.assign({visitor:function(n,r,o,i){return Zt.isNode&&P.isBuffer(n)?(this.append(r,n.toString("base64")),!1):i.defaultVisitor.apply(this,arguments)}},t))}function jN(e){return P.matchAll(/\w+|\[(\w*)]/g,e).map(t=>t[0]==="[]"?"":t[1]||t[0])}function BN(e){const t={},n=Object.keys(e);let r;const o=n.length;let i;for(r=0;r<o;r++)i=n[r],t[i]=e[i];return t}function Lg(e){function t(n,r,o,i){let s=n[i++];const l=Number.isFinite(+s),a=i>=n.length;return s=!s&&P.isArray(o)?o.length:s,a?(P.hasOwnProp(o,s)?o[s]=[o[s],r]:o[s]=r,!l):((!o[s]||!P.isObject(o[s]))&&(o[s]=[]),t(n,r,o[s],i)&&P.isArray(o[s])&&(o[s]=BN(o[s])),!l)}if(P.isFormData(e)&&P.isFunction(e.entries)){const n={};return P.forEachEntry(e,(r,o)=>{t(jN(r),o,n,0)}),n}return null}function zN(e,t,n){if(P.isString(e))try{return(t||JSON.parse)(e),P.trim(e)}catch(r){if(r.name!=="SyntaxError")throw r}return(n||JSON.stringify)(e)}const Uf={transitional:Ng,adapter:["xhr","http"],transformRequest:[function(t,n){const r=n.getContentType()||"",o=r.indexOf("application/json")>-1,i=P.isObject(t);if(i&&P.isHTMLForm(t)&&(t=new FormData(t)),P.isFormData(t))return o&&o?JSON.stringify(Lg(t)):t;if(P.isArrayBuffer(t)||P.isBuffer(t)||P.isStream(t)||P.isFile(t)||P.isBlob(t))return t;if(P.isArrayBufferView(t))return t.buffer;if(P.isURLSearchParams(t))return n.setContentType("application/x-www-form-urlencoded;charset=utf-8",!1),t.toString();let l;if(i){if(r.indexOf("application/x-www-form-urlencoded")>-1)return $N(t,this.formSerializer).toString();if((l=P.isFileList(t))||r.indexOf("multipart/form-data")>-1){const a=this.env&&this.env.FormData;return oa(l?{"files[]":t}:t,a&&new a,this.formSerializer)}}return i||o?(n.setContentType("application/json",!1),zN(t)):t}],transformResponse:[function(t){const n=this.transitional||Uf.transitional,r=n&&n.forcedJSONParsing,o=this.responseType==="json";if(t&&P.isString(t)&&(r&&!this.responseType||o)){const s=!(n&&n.silentJSONParsing)&&o;try{return JSON.parse(t)}catch(l){if(s)throw l.name==="SyntaxError"?me.from(l,me.ERR_BAD_RESPONSE,this,null,this.response):l}}return t}],timeout:0,xsrfCookieName:"XSRF-TOKEN",xsrfHeaderName:"X-XSRF-TOKEN",maxContentLength:-1,maxBodyLength:-1,env:{FormData:Zt.classes.FormData,Blob:Zt.classes.Blob},validateStatus:function(t){return t>=200&&t<300},headers:{common:{Accept:"application/json, text/plain, */*","Content-Type":void 0}}};P.forEach(["delete","get","head","post","put","patch"],e=>{Uf.headers[e]={}});const If=Uf,HN=P.toObjectSet(["age","authorization","content-length","content-type","etag","expires","from","host","if-modified-since","if-unmodified-since","last-modified","location","max-forwards","proxy-authorization","referer","retry-after","user-agent"]),WN=e=>{const t={};let n,r,o;return e&&e.split(`
-`).forEach(function(s){o=s.indexOf(":"),n=s.substring(0,o).trim().toLowerCase(),r=s.substring(o+1).trim(),!(!n||t[n]&&HN[n])&&(n==="set-cookie"?t[n]?t[n].push(r):t[n]=[r]:t[n]=t[n]?t[n]+", "+r:r)}),t},Yh=Symbol("internals");function Oo(e){return e&&String(e).trim().toLowerCase()}function Cs(e){return e===!1||e==null?e:P.isArray(e)?e.map(Cs):String(e)}function QN(e){const t=Object.create(null),n=/([^\s,;=]+)\s*(?:=\s*([^,;]+))?/g;let r;for(;r=n.exec(e);)t[r[1]]=r[2];return t}const qN=e=>/^[-_a-zA-Z0-9^`|~,!#$%&'*+.]+$/.test(e.trim());function Ha(e,t,n,r,o){if(P.isFunction(r))return r.call(this,t,n);if(o&&(t=n),!!P.isString(t)){if(P.isString(r))return t.indexOf(r)!==-1;if(P.isRegExp(r))return r.test(t)}}function KN(e){return e.trim().toLowerCase().replace(/([a-z\d])(\w*)/g,(t,n,r)=>n.toUpperCase()+r)}function GN(e,t){const n=P.toCamelCase(" "+t);["get","set","has"].forEach(r=>{Object.defineProperty(e,r+n,{value:function(o,i,s){return this[r].call(this,t,o,i,s)},configurable:!0})})}class ia{constructor(t){t&&this.set(t)}set(t,n,r){const o=this;function i(l,a,u){const f=Oo(a);if(!f)throw new Error("header name must be a non-empty string");const c=P.findKey(o,f);(!c||o[c]===void 0||u===!0||u===void 0&&o[c]!==!1)&&(o[c||a]=Cs(l))}const s=(l,a)=>P.forEach(l,(u,f)=>i(u,f,a));return P.isPlainObject(t)||t instanceof this.constructor?s(t,n):P.isString(t)&&(t=t.trim())&&!qN(t)?s(WN(t),n):t!=null&&i(n,t,r),this}get(t,n){if(t=Oo(t),t){const r=P.findKey(this,t);if(r){const o=this[r];if(!n)return o;if(n===!0)return QN(o);if(P.isFunction(n))return n.call(this,o,r);if(P.isRegExp(n))return n.exec(o);throw new TypeError("parser must be boolean|regexp|function")}}}has(t,n){if(t=Oo(t),t){const r=P.findKey(this,t);return!!(r&&this[r]!==void 0&&(!n||Ha(this,this[r],r,n)))}return!1}delete(t,n){const r=this;let o=!1;function i(s){if(s=Oo(s),s){const l=P.findKey(r,s);l&&(!n||Ha(r,r[l],l,n))&&(delete r[l],o=!0)}}return P.isArray(t)?t.forEach(i):i(t),o}clear(t){const n=Object.keys(this);let r=n.length,o=!1;for(;r--;){const i=n[r];(!t||Ha(this,this[i],i,t,!0))&&(delete this[i],o=!0)}return o}normalize(t){const n=this,r={};return P.forEach(this,(o,i)=>{const s=P.findKey(r,i);if(s){n[s]=Cs(o),delete n[i];return}const l=t?KN(i):String(i).trim();l!==i&&delete n[i],n[l]=Cs(o),r[l]=!0}),this}concat(...t){return this.constructor.concat(this,...t)}toJSON(t){const n=Object.create(null);return P.forEach(this,(r,o)=>{r!=null&&r!==!1&&(n[o]=t&&P.isArray(r)?r.join(", "):r)}),n}[Symbol.iterator](){return Object.entries(this.toJSON())[Symbol.iterator]()}toString(){return Object.entries(this.toJSON()).map(([t,n])=>t+": "+n).join(`
-`)}get[Symbol.toStringTag](){return"AxiosHeaders"}static from(t){return t instanceof this?t:new this(t)}static concat(t,...n){const r=new this(t);return n.forEach(o=>r.set(o)),r}static accessor(t){const r=(this[Yh]=this[Yh]={accessors:{}}).accessors,o=this.prototype;function i(s){const l=Oo(s);r[l]||(GN(o,s),r[l]=!0)}return P.isArray(t)?t.forEach(i):i(t),this}}ia.accessor(["Content-Type","Content-Length","Accept","Accept-Encoding","User-Agent","Authorization"]);P.reduceDescriptors(ia.prototype,({value:e},t)=>{let n=t[0].toUpperCase()+t.slice(1);return{get:()=>e,set(r){this[n]=r}}});P.freezeMethods(ia);const hn=ia;function Wa(e,t){const n=this||If,r=t||n,o=hn.from(r.headers);let i=r.data;return P.forEach(e,function(l){i=l.call(n,i,o.normalize(),t?t.status:void 0)}),o.normalize(),i}function Pg(e){return!!(e&&e.__CANCEL__)}function bi(e,t,n){me.call(this,e??"canceled",me.ERR_CANCELED,t,n),this.name="CanceledError"}P.inherits(bi,me,{__CANCEL__:!0});function YN(e,t,n){const r=n.config.validateStatus;!n.status||!r||r(n.status)?e(n):t(new me("Request failed with status code "+n.status,[me.ERR_BAD_REQUEST,me.ERR_BAD_RESPONSE][Math.floor(n.status/100)-4],n.config,n.request,n))}const JN=Zt.hasStandardBrowserEnv?{write(e,t,n,r,o,i){const s=[e+"="+encodeURIComponent(t)];P.isNumber(n)&&s.push("expires="+new Date(n).toGMTString()),P.isString(r)&&s.push("path="+r),P.isString(o)&&s.push("domain="+o),i===!0&&s.push("secure"),document.cookie=s.join("; ")},read(e){const t=document.cookie.match(new RegExp("(^|;\\s*)("+e+")=([^;]*)"));return t?decodeURIComponent(t[3]):null},remove(e){this.write(e,"",Date.now()-864e5)}}:{write(){},read(){return null},remove(){}};function XN(e){return/^([a-z][a-z\d+\-.]*:)?\/\//i.test(e)}function ZN(e,t){return t?e.replace(/\/+$/,"")+"/"+t.replace(/^\/+/,""):e}function Og(e,t){return e&&!XN(t)?ZN(e,t):t}const eA=Zt.hasStandardBrowserEnv?function(){const t=/(msie|trident)/i.test(navigator.userAgent),n=document.createElement("a");let r;function o(i){let s=i;return t&&(n.setAttribute("href",s),s=n.href),n.setAttribute("href",s),{href:n.href,protocol:n.protocol?n.protocol.replace(/:$/,""):"",host:n.host,search:n.search?n.search.replace(/^\?/,""):"",hash:n.hash?n.hash.replace(/^#/,""):"",hostname:n.hostname,port:n.port,pathname:n.pathname.charAt(0)==="/"?n.pathname:"/"+n.pathname}}return r=o(window.location.href),function(s){const l=P.isString(s)?o(s):s;return l.protocol===r.protocol&&l.host===r.host}}():function(){return function(){return!0}}();function tA(e){const t=/^([-+\w]{1,25})(:?\/\/|:)/.exec(e);return t&&t[1]||""}function nA(e,t){e=e||10;const n=new Array(e),r=new Array(e);let o=0,i=0,s;return t=t!==void 0?t:1e3,function(a){const u=Date.now(),f=r[i];s||(s=u),n[o]=a,r[o]=u;let c=i,h=0;for(;c!==o;)h+=n[c++],c=c%e;if(o=(o+1)%e,o===i&&(i=(i+1)%e),u-s<t)return;const S=f&&u-f;return S?Math.round(h*1e3/S):void 0}}function Jh(e,t){let n=0;const r=nA(50,250);return o=>{const i=o.loaded,s=o.lengthComputable?o.total:void 0,l=i-n,a=r(l),u=i<=s;n=i;const f={loaded:i,total:s,progress:s?i/s:void 0,bytes:l,rate:a||void 0,estimated:a&&s&&u?(s-i)/a:void 0,event:o};f[t?"download":"upload"]=!0,e(f)}}const rA=typeof XMLHttpRequest<"u",oA=rA&&function(e){return new Promise(function(n,r){let o=e.data;const i=hn.from(e.headers).normalize();let{responseType:s,withXSRFToken:l}=e,a;function u(){e.cancelToken&&e.cancelToken.unsubscribe(a),e.signal&&e.signal.removeEventListener("abort",a)}let f;if(P.isFormData(o)){if(Zt.hasStandardBrowserEnv||Zt.hasStandardBrowserWebWorkerEnv)i.setContentType(!1);else if((f=i.getContentType())!==!1){const[v,...T]=f?f.split(";").map(m=>m.trim()).filter(Boolean):[];i.setContentType([v||"multipart/form-data",...T].join("; "))}}let c=new XMLHttpRequest;if(e.auth){const v=e.auth.username||"",T=e.auth.password?unescape(encodeURIComponent(e.auth.password)):"";i.set("Authorization","Basic "+btoa(v+":"+T))}const h=Og(e.baseURL,e.url);c.open(e.method.toUpperCase(),kg(h,e.params,e.paramsSerializer),!0),c.timeout=e.timeout;function S(){if(!c)return;const v=hn.from("getAllResponseHeaders"in c&&c.getAllResponseHeaders()),m={data:!s||s==="text"||s==="json"?c.responseText:c.response,status:c.status,statusText:c.statusText,headers:v,config:e,request:c};YN(function(y){n(y),u()},function(y){r(y),u()},m),c=null}if("onloadend"in c?c.onloadend=S:c.onreadystatechange=function(){!c||c.readyState!==4||c.status===0&&!(c.responseURL&&c.responseURL.indexOf("file:")===0)||setTimeout(S)},c.onabort=function(){c&&(r(new me("Request aborted",me.ECONNABORTED,e,c)),c=null)},c.onerror=function(){r(new me("Network Error",me.ERR_NETWORK,e,c)),c=null},c.ontimeout=function(){let T=e.timeout?"timeout of "+e.timeout+"ms exceeded":"timeout exceeded";const m=e.transitional||Ng;e.timeoutErrorMessage&&(T=e.timeoutErrorMessage),r(new me(T,m.clarifyTimeoutError?me.ETIMEDOUT:me.ECONNABORTED,e,c)),c=null},Zt.hasStandardBrowserEnv&&(l&&P.isFunction(l)&&(l=l(e)),l||l!==!1&&eA(h))){const v=e.xsrfHeaderName&&e.xsrfCookieName&&JN.read(e.xsrfCookieName);v&&i.set(e.xsrfHeaderName,v)}o===void 0&&i.setContentType(null),"setRequestHeader"in c&&P.forEach(i.toJSON(),function(T,m){c.setRequestHeader(m,T)}),P.isUndefined(e.withCredentials)||(c.withCredentials=!!e.withCredentials),s&&s!=="json"&&(c.responseType=e.responseType),typeof e.onDownloadProgress=="function"&&c.addEventListener("progress",Jh(e.onDownloadProgress,!0)),typeof e.onUploadProgress=="function"&&c.upload&&c.upload.addEventListener("progress",Jh(e.onUploadProgress)),(e.cancelToken||e.signal)&&(a=v=>{c&&(r(!v||v.type?new bi(null,e,c):v),c.abort(),c=null)},e.cancelToken&&e.cancelToken.subscribe(a),e.signal&&(e.signal.aborted?a():e.signal.addEventListener("abort",a)));const p=tA(h);if(p&&Zt.protocols.indexOf(p)===-1){r(new me("Unsupported protocol "+p+":",me.ERR_BAD_REQUEST,e));return}c.send(o||null)})},cc={http:NN,xhr:oA};P.forEach(cc,(e,t)=>{if(e){try{Object.defineProperty(e,"name",{value:t})}catch{}Object.defineProperty(e,"adapterName",{value:t})}});const Xh=e=>`- ${e}`,iA=e=>P.isFunction(e)||e===null||e===!1,bg={getAdapter:e=>{e=P.isArray(e)?e:[e];const{length:t}=e;let n,r;const o={};for(let i=0;i<t;i++){n=e[i];let s;if(r=n,!iA(n)&&(r=cc[(s=String(n)).toLowerCase()],r===void 0))throw new me(`Unknown adapter '${s}'`);if(r)break;o[s||"#"+i]=r}if(!r){const i=Object.entries(o).map(([l,a])=>`adapter ${l} `+(a===!1?"is not supported by the environment":"is not available in the build"));let s=t?i.length>1?`since :
-`+i.map(Xh).join(`
-`):" "+Xh(i[0]):"as no adapter specified";throw new me("There is no suitable adapter to dispatch the request "+s,"ERR_NOT_SUPPORT")}return r},adapters:cc};function Qa(e){if(e.cancelToken&&e.cancelToken.throwIfRequested(),e.signal&&e.signal.aborted)throw new bi(null,e)}function Zh(e){return Qa(e),e.headers=hn.from(e.headers),e.data=Wa.call(e,e.transformRequest),["post","put","patch"].indexOf(e.method)!==-1&&e.headers.setContentType("application/x-www-form-urlencoded",!1),bg.getAdapter(e.adapter||If.adapter)(e).then(function(r){return Qa(e),r.data=Wa.call(e,e.transformResponse,r),r.headers=hn.from(r.headers),r},function(r){return Pg(r)||(Qa(e),r&&r.response&&(r.response.data=Wa.call(e,e.transformResponse,r.response),r.response.headers=hn.from(r.response.headers))),Promise.reject(r)})}const ep=e=>e instanceof hn?e.toJSON():e;function io(e,t){t=t||{};const n={};function r(u,f,c){return P.isPlainObject(u)&&P.isPlainObject(f)?P.merge.call({caseless:c},u,f):P.isPlainObject(f)?P.merge({},f):P.isArray(f)?f.slice():f}function o(u,f,c){if(P.isUndefined(f)){if(!P.isUndefined(u))return r(void 0,u,c)}else return r(u,f,c)}function i(u,f){if(!P.isUndefined(f))return r(void 0,f)}function s(u,f){if(P.isUndefined(f)){if(!P.isUndefined(u))return r(void 0,u)}else return r(void 0,f)}function l(u,f,c){if(c in t)return r(u,f);if(c in e)return r(void 0,u)}const a={url:i,method:i,data:i,baseURL:s,transformRequest:s,transformResponse:s,paramsSerializer:s,timeout:s,timeoutMessage:s,withCredentials:s,withXSRFToken:s,adapter:s,responseType:s,xsrfCookieName:s,xsrfHeaderName:s,onUploadProgress:s,onDownloadProgress:s,decompress:s,maxContentLength:s,maxBodyLength:s,beforeRedirect:s,transport:s,httpAgent:s,httpsAgent:s,cancelToken:s,socketPath:s,responseEncoding:s,validateStatus:l,headers:(u,f)=>o(ep(u),ep(f),!0)};return P.forEach(Object.keys(Object.assign({},e,t)),function(f){const c=a[f]||o,h=c(e[f],t[f],f);P.isUndefined(h)&&c!==l||(n[f]=h)}),n}const Fg="1.6.2",Vf={};["object","boolean","number","function","string","symbol"].forEach((e,t)=>{Vf[e]=function(r){return typeof r===e||"a"+(t<1?"n ":" ")+e}});const tp={};Vf.transitional=function(t,n,r){function o(i,s){return"[Axios v"+Fg+"] Transitional option '"+i+"'"+s+(r?". "+r:"")}return(i,s,l)=>{if(t===!1)throw new me(o(s," has been removed"+(n?" in "+n:"")),me.ERR_DEPRECATED);return n&&!tp[s]&&(tp[s]=!0,console.warn(o(s," has been deprecated since v"+n+" and will be removed in the near future"))),t?t(i,s,l):!0}};function sA(e,t,n){if(typeof e!="object")throw new me("options must be an object",me.ERR_BAD_OPTION_VALUE);const r=Object.keys(e);let o=r.length;for(;o-- >0;){const i=r[o],s=t[i];if(s){const l=e[i],a=l===void 0||s(l,i,e);if(a!==!0)throw new me("option "+i+" must be "+a,me.ERR_BAD_OPTION_VALUE);continue}if(n!==!0)throw new me("Unknown option "+i,me.ERR_BAD_OPTION)}}const fc={assertOptions:sA,validators:Vf},Rn=fc.validators;class al{constructor(t){this.defaults=t,this.interceptors={request:new Gh,response:new Gh}}request(t,n){typeof t=="string"?(n=n||{},n.url=t):n=t||{},n=io(this.defaults,n);const{transitional:r,paramsSerializer:o,headers:i}=n;r!==void 0&&fc.assertOptions(r,{silentJSONParsing:Rn.transitional(Rn.boolean),forcedJSONParsing:Rn.transitional(Rn.boolean),clarifyTimeoutError:Rn.transitional(Rn.boolean)},!1),o!=null&&(P.isFunction(o)?n.paramsSerializer={serialize:o}:fc.assertOptions(o,{encode:Rn.function,serialize:Rn.function},!0)),n.method=(n.method||this.defaults.method||"get").toLowerCase();let s=i&&P.merge(i.common,i[n.method]);i&&P.forEach(["delete","get","head","post","put","patch","common"],p=>{delete i[p]}),n.headers=hn.concat(s,i);const l=[];let a=!0;this.interceptors.request.forEach(function(v){typeof v.runWhen=="function"&&v.runWhen(n)===!1||(a=a&&v.synchronous,l.unshift(v.fulfilled,v.rejected))});const u=[];this.interceptors.response.forEach(function(v){u.push(v.fulfilled,v.rejected)});let f,c=0,h;if(!a){const p=[Zh.bind(this),void 0];for(p.unshift.apply(p,l),p.push.apply(p,u),h=p.length,f=Promise.resolve(n);c<h;)f=f.then(p[c++],p[c++]);return f}h=l.length;let S=n;for(c=0;c<h;){const p=l[c++],v=l[c++];try{S=p(S)}catch(T){v.call(this,T);break}}try{f=Zh.call(this,S)}catch(p){return Promise.reject(p)}for(c=0,h=u.length;c<h;)f=f.then(u[c++],u[c++]);return f}getUri(t){t=io(this.defaults,t);const n=Og(t.baseURL,t.url);return kg(n,t.params,t.paramsSerializer)}}P.forEach(["delete","get","head","options"],function(t){al.prototype[t]=function(n,r){return this.request(io(r||{},{method:t,url:n,data:(r||{}).data}))}});P.forEach(["post","put","patch"],function(t){function n(r){return function(i,s,l){return this.request(io(l||{},{method:t,headers:r?{"Content-Type":"multipart/form-data"}:{},url:i,data:s}))}}al.prototype[t]=n(),al.prototype[t+"Form"]=n(!0)});const xs=al;class $f{constructor(t){if(typeof t!="function")throw new TypeError("executor must be a function.");let n;this.promise=new Promise(function(i){n=i});const r=this;this.promise.then(o=>{if(!r._listeners)return;let i=r._listeners.length;for(;i-- >0;)r._listeners[i](o);r._listeners=null}),this.promise.then=o=>{let i;const s=new Promise(l=>{r.subscribe(l),i=l}).then(o);return s.cancel=function(){r.unsubscribe(i)},s},t(function(i,s,l){r.reason||(r.reason=new bi(i,s,l),n(r.reason))})}throwIfRequested(){if(this.reason)throw this.reason}subscribe(t){if(this.reason){t(this.reason);return}this._listeners?this._listeners.push(t):this._listeners=[t]}unsubscribe(t){if(!this._listeners)return;const n=this._listeners.indexOf(t);n!==-1&&this._listeners.splice(n,1)}static source(){let t;return{token:new $f(function(o){t=o}),cancel:t}}}const lA=$f;function aA(e){return function(n){return e.apply(null,n)}}function uA(e){return P.isObject(e)&&e.isAxiosError===!0}const dc={Continue:100,SwitchingProtocols:101,Processing:102,EarlyHints:103,Ok:200,Created:201,Accepted:202,NonAuthoritativeInformation:203,NoContent:204,ResetContent:205,PartialContent:206,MultiStatus:207,AlreadyReported:208,ImUsed:226,MultipleChoices:300,MovedPermanently:301,Found:302,SeeOther:303,NotModified:304,UseProxy:305,Unused:306,TemporaryRedirect:307,PermanentRedirect:308,BadRequest:400,Unauthorized:401,PaymentRequired:402,Forbidden:403,NotFound:404,MethodNotAllowed:405,NotAcceptable:406,ProxyAuthenticationRequired:407,RequestTimeout:408,Conflict:409,Gone:410,LengthRequired:411,PreconditionFailed:412,PayloadTooLarge:413,UriTooLong:414,UnsupportedMediaType:415,RangeNotSatisfiable:416,ExpectationFailed:417,ImATeapot:418,MisdirectedRequest:421,UnprocessableEntity:422,Locked:423,FailedDependency:424,TooEarly:425,UpgradeRequired:426,PreconditionRequired:428,TooManyRequests:429,RequestHeaderFieldsTooLarge:431,UnavailableForLegalReasons:451,InternalServerError:500,NotImplemented:501,BadGateway:502,ServiceUnavailable:503,GatewayTimeout:504,HttpVersionNotSupported:505,VariantAlsoNegotiates:506,InsufficientStorage:507,LoopDetected:508,NotExtended:510,NetworkAuthenticationRequired:511};Object.entries(dc).forEach(([e,t])=>{dc[t]=e});const cA=dc;function Dg(e){const t=new xs(e),n=mg(xs.prototype.request,t);return P.extend(n,xs.prototype,t,{allOwnKeys:!0}),P.extend(n,t,null,{allOwnKeys:!0}),n.create=function(o){return Dg(io(e,o))},n}const We=Dg(If);We.Axios=xs;We.CanceledError=bi;We.CancelToken=lA;We.isCancel=Pg;We.VERSION=Fg;We.toFormData=oa;We.AxiosError=me;We.Cancel=We.CanceledError;We.all=function(t){return Promise.all(t)};We.spread=aA;We.isAxiosError=uA;We.mergeConfig=io;We.AxiosHeaders=hn;We.formToJSON=e=>Lg(P.isHTMLForm(e)?new FormData(e):e);We.getAdapter=bg.getAdapter;We.HttpStatusCode=cA;We.default=We;const Mg=We,Zo=Mg.create({baseURL:"/services/japps",headers:{"Content-Type":"application/json"},params:{token:"13748105b2e64b9082e10f802b04f523"}});Lf({key:"currentUser",default:void 0});const jf=Lf({key:"currentJhData",default:{admin_access:!1,base_url:"/hub",options_form:!1,prefix:"/",user:"",xsrf_token:""}}),Fi=Lf({key:"currentNotification",default:void 0}),so=({id:e,type:t="button",variant:n="primary",className:r,children:o,...i})=>{const s=Yn("btn rounded border-2 font-bold whitespace-nowrap",{"btn-primary":n==="primary","btn-secondary":n==="secondary"},r);return C.jsx("button",{id:e,type:t,className:s,...i,children:o})};function fA({title:e,titleId:t,...n},r){return U.createElement("svg",Object.assign({xmlns:"http://www.w3.org/2000/svg",viewBox:"0 0 24 24",fill:"currentColor","aria-hidden":"true",ref:r,"aria-labelledby":t},n),e?U.createElement("title",{id:t},e):null,U.createElement("path",{fillRule:"evenodd",d:"M4.5 12a1.5 1.5 0 113 0 1.5 1.5 0 01-3 0zm6 0a1.5 1.5 0 113 0 1.5 1.5 0 01-3 0zm6 0a1.5 1.5 0 113 0 1.5 1.5 0 01-3 0z",clipRule:"evenodd"}))}const dA=U.forwardRef(fA),hA=dA;function pA({title:e,titleId:t,...n},r){return U.createElement("svg",Object.assign({xmlns:"http://www.w3.org/2000/svg",viewBox:"0 0 24 24",fill:"currentColor","aria-hidden":"true",ref:r,"aria-labelledby":t},n),e?U.createElement("title",{id:t},e):null,U.createElement("path",{fillRule:"evenodd",d:"M5.47 5.47a.75.75 0 011.06 0L12 10.94l5.47-5.47a.75.75 0 111.06 1.06L13.06 12l5.47 5.47a.75.75 0 11-1.06 1.06L12 13.06l-5.47 5.47a.75.75 0 01-1.06-1.06L10.94 12 5.47 6.53a.75.75 0 010-1.06z",clipRule:"evenodd"}))}const mA=U.forwardRef(pA),yA=mA,vA=({id:e,items:t})=>{const n=ne.useRef(null);return C.jsxs("div",{className:"context-menu-container",id:e,tabIndex:0,ref:n,children:[C.jsx(hA,{}),C.jsx("ul",{className:"context-menu",children:t.filter(r=>r.visible).map(r=>C.jsx("li",{children:C.jsx("a",{className:r.disabled?"disabled":"",onClick:o=>{var i;r.onClick&&!r.disabled&&(r.onClick(o),(i=n.current)==null||i.blur())},children:r.title})},`context-menu-item-${r.id}`))})]})},hc=({title:e,setIsOpen:t,body:n,footer:r})=>C.jsxs(C.Fragment,{children:[C.jsx("div",{className:"modal-overlay",onClick:()=>t(!1)}),C.jsx("div",{className:"modal-container",children:C.jsxs("div",{className:"modal-main",children:[C.jsxs("div",{className:"modal-heading",children:[C.jsx("h5",{className:"modal-title",children:e}),C.jsx("button",{className:"modal-close-btn",onClick:()=>t(!1),children:C.jsx(yA,{})})]}),C.jsx("div",{className:"modal-body",children:n}),r?C.jsx("div",{className:"modal-footer",children:r}):C.jsx(C.Fragment,{})]})})]}),gA=({id:e,children:t,className:n})=>{const r=Yn("tag",n);return C.jsx("span",{id:e,className:r,children:t})},SA=({id:e,title:t,description:n,thumbnail:r,framework:o,url:i,ready:s=!1})=>{const l=ea(),[a,u]=U.useState(!1),[,f]=Hn(Fi),[c,h]=U.useState(!1),[S,p]=U.useState(!1),v=async({id:_})=>await Zo.delete(`/server/${_}`),{mutate:T}=lc({mutationFn:v,retry:1}),m=()=>{u(!0),T({id:e},{onSuccess:async()=>{u(!1),h(!1),l.invalidateQueries(["app-state"])},onError:async _=>{u(!1),f(_.message)}})},d=[{id:"start",title:"Start",onClick:()=>window.location.assign(i),disabled:s,visible:!1},{id:"stop",title:"Stop",onClick:()=>console.log(`Stopping app with id: ${e}`),disabled:!s,visible:!1},{id:"edit",title:"Edit",onClick:()=>p(!0),visible:!0},{id:"delete",title:"Delete",onClick:()=>h(!0),visible:!0}],y=C.jsxs(C.Fragment,{children:[C.jsxs("p",{className:"w-[400px] mb-6",children:["Are you sure you want to delete ",C.jsx("b",{children:t}),"? This action is permanent and cannot be reversed."]}),C.jsxs(Xg,{children:[C.jsx(so,{id:"cancel-btn",variant:"secondary",onClick:()=>h(!1),children:"Cancel"}),C.jsx(so,{id:"delete-btn",variant:"primary",onClick:()=>m(),disabled:a,children:"Delete"})]})]});return C.jsxs("div",{className:"card",id:`card-${e}`,tabIndex:0,children:[C.jsxs("div",{className:"card-header-media",children:[C.jsxs("div",{className:"card-header-menu",children:[C.jsx(vA,{id:`card-menu-${e}`,items:d}),c&&C.jsx(hc,{title:`Delete ${t}`,setIsOpen:h,body:y}),S&&C.jsx(hc,{title:`Edit ${t}`,setIsOpen:p,body:C.jsx(Jg,{id:e,onCancel:()=>p(!1),onSubmit:()=>p(!1)})})]}),C.jsx("div",{className:"card-header-img",children:r?C.jsx("img",{src:r,alt:"App thumb"}):void 0})]}),C.jsx("div",{className:"card-header",children:C.jsx("h3",{className:"font-bold",children:C.jsx("a",{href:i,children:t})})}),C.jsx("div",{className:"card-body",children:C.jsx("p",{className:"text-sm",children:n})}),C.jsx("div",{className:"card-footer",children:C.jsx(gA,{id:`tag-${e}`,children:o})})]})},wA="This field is required.",np={required:wA},Bf=()=>window.jhdata,_A=(e,t)=>{var r;const n=[];for(const o in e)if(Object.hasOwnProperty.call(e,o)){const i=e[o];i.display===!0&&n.push({name:i.info.name,url:(r=i.info.url)==null?void 0:r.replace("[USER]",t),external:i.info.external})}return n},RA=(e,t)=>{var r;const n=[];for(const o in e)if(Object.hasOwnProperty.call(e,o)){const i=e[o];if((r=i.user_options)!=null&&r.jhub_app){const s=i.user_options;n.push({id:s.name,name:s.display_name,description:s.description,framework:EA(s.framework),url:i.url,thumbnail:s.thumbnail,shared:!1,ready:s.ready})}}return t.toLowerCase()==="shared"?n.filter(o=>o.shared===!0):n.filter(o=>o.shared===!1)},EA=e=>e.charAt(0).toUpperCase()+e.slice(1);var Di=e=>e.type==="checkbox",Ur=e=>e instanceof Date,at=e=>e==null;const Ug=e=>typeof e=="object";var Ke=e=>!at(e)&&!Array.isArray(e)&&Ug(e)&&!Ur(e),Ig=e=>Ke(e)&&e.target?Di(e.target)?e.target.checked:e.target.value:e,CA=e=>e.substring(0,e.search(/\.\d+(\.|$)/))||e,Vg=(e,t)=>e.has(CA(t)),xA=e=>{const t=e.constructor&&e.constructor.prototype;return Ke(t)&&t.hasOwnProperty("isPrototypeOf")},zf=typeof window<"u"&&typeof window.HTMLElement<"u"&&typeof document<"u";function Bt(e){let t;const n=Array.isArray(e);if(e instanceof Date)t=new Date(e);else if(e instanceof Set)t=new Set(e);else if(!(zf&&(e instanceof Blob||e instanceof FileList))&&(n||Ke(e)))if(t=n?[]:{},!n&&!xA(e))t=e;else for(const r in e)e.hasOwnProperty(r)&&(t[r]=Bt(e[r]));else return e;return t}var Mi=e=>Array.isArray(e)?e.filter(Boolean):[],je=e=>e===void 0,W=(e,t,n)=>{if(!t||!Ke(e))return n;const r=Mi(t.split(/[,[\].]+?/)).reduce((o,i)=>at(o)?o:o[i],e);return je(r)||r===e?je(e[t])?n:e[t]:r},On=e=>typeof e=="boolean";const ul={BLUR:"blur",FOCUS_OUT:"focusout",CHANGE:"change"},Wt={onBlur:"onBlur",onChange:"onChange",onSubmit:"onSubmit",onTouched:"onTouched",all:"all"},ln={max:"max",min:"min",maxLength:"maxLength",minLength:"minLength",pattern:"pattern",required:"required",validate:"validate"},TA=ne.createContext(null),Hf=()=>ne.useContext(TA);var $g=(e,t,n,r=!0)=>{const o={defaultValues:t._defaultValues};for(const i in e)Object.defineProperty(o,i,{get:()=>{const s=i;return t._proxyFormState[s]!==Wt.all&&(t._proxyFormState[s]=!r||Wt.all),n&&(n[s]=!0),e[s]}});return o},Lt=e=>Ke(e)&&!Object.keys(e).length,jg=(e,t,n,r)=>{n(e);const{name:o,...i}=e;return Lt(i)||Object.keys(i).length>=Object.keys(t).length||Object.keys(i).find(s=>t[s]===(!r||Wt.all))},Ts=e=>Array.isArray(e)?e:[e],Bg=(e,t,n)=>!e||!t||e===t||Ts(e).some(r=>r&&(n?r===t:r.startsWith(t)||t.startsWith(r)));function Wf(e){const t=ne.useRef(e);t.current=e,ne.useEffect(()=>{const n=!e.disabled&&t.current.subject&&t.current.subject.subscribe({next:t.current.next});return()=>{n&&n.unsubscribe()}},[e.disabled])}function kA(e){const t=Hf(),{control:n=t.control,disabled:r,name:o,exact:i}=e||{},[s,l]=ne.useState(n._formState),a=ne.useRef(!0),u=ne.useRef({isDirty:!1,isLoading:!1,dirtyFields:!1,touchedFields:!1,isValidating:!1,isValid:!1,errors:!1}),f=ne.useRef(o);return f.current=o,Wf({disabled:r,next:c=>a.current&&Bg(f.current,c.name,i)&&jg(c,u.current,n._updateFormState)&&l({...n._formState,...c}),subject:n._subjects.state}),ne.useEffect(()=>(a.current=!0,u.current.isValid&&n._updateValid(!0),()=>{a.current=!1}),[n]),$g(s,n,u.current,!1)}var en=e=>typeof e=="string",zg=(e,t,n,r,o)=>en(e)?(r&&t.watch.add(e),W(n,e,o)):Array.isArray(e)?e.map(i=>(r&&t.watch.add(i),W(n,i))):(r&&(t.watchAll=!0),n);function NA(e){const t=Hf(),{control:n=t.control,name:r,defaultValue:o,disabled:i,exact:s}=e||{},l=ne.useRef(r);l.current=r,Wf({disabled:i,subject:n._subjects.values,next:f=>{Bg(l.current,f.name,s)&&u(Bt(zg(l.current,n._names,f.values||n._formValues,!1,o)))}});const[a,u]=ne.useState(n._getWatch(r,o));return ne.useEffect(()=>n._removeUnmounted()),a}var Qf=e=>/^\w*$/.test(e),Hg=e=>Mi(e.replace(/["|']|\]/g,"").split(/\.|\[/));function ke(e,t,n){let r=-1;const o=Qf(t)?[t]:Hg(t),i=o.length,s=i-1;for(;++r<i;){const l=o[r];let a=n;if(r!==s){const u=e[l];a=Ke(u)||Array.isArray(u)?u:isNaN(+o[r+1])?{}:[]}e[l]=a,e=e[l]}return e}function AA(e){const t=Hf(),{name:n,disabled:r,control:o=t.control,shouldUnregister:i}=e,s=Vg(o._names.array,n),l=NA({control:o,name:n,defaultValue:W(o._formValues,n,W(o._defaultValues,n,e.defaultValue)),exact:!0}),a=kA({control:o,name:n}),u=ne.useRef(o.register(n,{...e.rules,value:l}));return u.current=o.register(n,e.rules),ne.useEffect(()=>{const f=o._options.shouldUnregister||i,c=(h,S)=>{const p=W(o._fields,h);p&&(p._f.mount=S)};if(c(n,!0),f){const h=Bt(W(o._options.defaultValues,n));ke(o._defaultValues,n,h),je(W(o._formValues,n))&&ke(o._formValues,n,h)}return()=>{(s?f&&!o._state.action:f)?o.unregister(n):c(n,!1)}},[n,o,s,i]),ne.useEffect(()=>{W(o._fields,n)&&o._updateDisabledField({disabled:r,fields:o._fields,name:n})},[r,n,o]),{field:{name:n,value:l,...On(r)?{disabled:r}:{},onChange:ne.useCallback(f=>u.current.onChange({target:{value:Ig(f),name:n},type:ul.CHANGE}),[n]),onBlur:ne.useCallback(()=>u.current.onBlur({target:{value:W(o._formValues,n),name:n},type:ul.BLUR}),[n,o]),ref:f=>{const c=W(o._fields,n);c&&f&&(c._f.ref={focus:()=>f.focus(),select:()=>f.select(),setCustomValidity:h=>f.setCustomValidity(h),reportValidity:()=>f.reportValidity()})}},formState:a,fieldState:Object.defineProperties({},{invalid:{enumerable:!0,get:()=>!!W(a.errors,n)},isDirty:{enumerable:!0,get:()=>!!W(a.dirtyFields,n)},isTouched:{enumerable:!0,get:()=>!!W(a.touchedFields,n)},error:{enumerable:!0,get:()=>W(a.errors,n)}})}}const _r=e=>e.render(AA(e));var LA=(e,t,n,r,o)=>t?{...n[e],types:{...n[e]&&n[e].types?n[e].types:{},[r]:o||!0}}:{};const pc=(e,t,n)=>{for(const r of n||Object.keys(e)){const o=W(e,r);if(o){const{_f:i,...s}=o;if(i&&t(i.name)){if(i.ref.focus){i.ref.focus();break}else if(i.refs&&i.refs[0].focus){i.refs[0].focus();break}}else Ke(s)&&pc(s,t)}}};var rp=e=>({isOnSubmit:!e||e===Wt.onSubmit,isOnBlur:e===Wt.onBlur,isOnChange:e===Wt.onChange,isOnAll:e===Wt.all,isOnTouch:e===Wt.onTouched}),op=(e,t,n)=>!n&&(t.watchAll||t.watch.has(e)||[...t.watch].some(r=>e.startsWith(r)&&/^\.\w+/.test(e.slice(r.length)))),PA=(e,t,n)=>{const r=Mi(W(e,n));return ke(r,"root",t[n]),ke(e,n,r),e},qf=e=>e.type==="file",bn=e=>typeof e=="function",cl=e=>{if(!zf)return!1;const t=e?e.ownerDocument:0;return e instanceof(t&&t.defaultView?t.defaultView.HTMLElement:HTMLElement)},ks=e=>en(e),Kf=e=>e.type==="radio",fl=e=>e instanceof RegExp;const ip={value:!1,isValid:!1},sp={value:!0,isValid:!0};var Wg=e=>{if(Array.isArray(e)){if(e.length>1){const t=e.filter(n=>n&&n.checked&&!n.disabled).map(n=>n.value);return{value:t,isValid:!!t.length}}return e[0].checked&&!e[0].disabled?e[0].attributes&&!je(e[0].attributes.value)?je(e[0].value)||e[0].value===""?sp:{value:e[0].value,isValid:!0}:sp:ip}return ip};const lp={isValid:!1,value:null};var Qg=e=>Array.isArray(e)?e.reduce((t,n)=>n&&n.checked&&!n.disabled?{isValid:!0,value:n.value}:t,lp):lp;function ap(e,t,n="validate"){if(ks(e)||Array.isArray(e)&&e.every(ks)||On(e)&&!e)return{type:n,message:ks(e)?e:"",ref:t}}var Rr=e=>Ke(e)&&!fl(e)?e:{value:e,message:""},up=async(e,t,n,r,o)=>{const{ref:i,refs:s,required:l,maxLength:a,minLength:u,min:f,max:c,pattern:h,validate:S,name:p,valueAsNumber:v,mount:T,disabled:m}=e._f,d=W(t,p);if(!T||m)return{};const y=s?s[0]:i,_=ae=>{r&&y.reportValidity&&(y.setCustomValidity(On(ae)?"":ae||""),y.reportValidity())},A={},k=Kf(i),x=Di(i),O=k||x,J=(v||qf(i))&&je(i.value)&&je(d)||cl(i)&&i.value===""||d===""||Array.isArray(d)&&!d.length,G=LA.bind(null,p,n,A),ye=(ae,B,ie,pe=ln.maxLength,ge=ln.minLength)=>{const _e=ae?B:ie;A[p]={type:ae?pe:ge,message:_e,ref:i,...G(ae?pe:ge,_e)}};if(o?!Array.isArray(d)||!d.length:l&&(!O&&(J||at(d))||On(d)&&!d||x&&!Wg(s).isValid||k&&!Qg(s).isValid)){const{value:ae,message:B}=ks(l)?{value:!!l,message:l}:Rr(l);if(ae&&(A[p]={type:ln.required,message:B,ref:y,...G(ln.required,B)},!n))return _(B),A}if(!J&&(!at(f)||!at(c))){let ae,B;const ie=Rr(c),pe=Rr(f);if(!at(d)&&!isNaN(d)){const ge=i.valueAsNumber||d&&+d;at(ie.value)||(ae=ge>ie.value),at(pe.value)||(B=ge<pe.value)}else{const ge=i.valueAsDate||new Date(d),_e=te=>new Date(new Date().toDateString()+" "+te),V=i.type=="time",X=i.type=="week";en(ie.value)&&d&&(ae=V?_e(d)>_e(ie.value):X?d>ie.value:ge>new Date(ie.value)),en(pe.value)&&d&&(B=V?_e(d)<_e(pe.value):X?d<pe.value:ge<new Date(pe.value))}if((ae||B)&&(ye(!!ae,ie.message,pe.message,ln.max,ln.min),!n))return _(A[p].message),A}if((a||u)&&!J&&(en(d)||o&&Array.isArray(d))){const ae=Rr(a),B=Rr(u),ie=!at(ae.value)&&d.length>+ae.value,pe=!at(B.value)&&d.length<+B.value;if((ie||pe)&&(ye(ie,ae.message,B.message),!n))return _(A[p].message),A}if(h&&!J&&en(d)){const{value:ae,message:B}=Rr(h);if(fl(ae)&&!d.match(ae)&&(A[p]={type:ln.pattern,message:B,ref:i,...G(ln.pattern,B)},!n))return _(B),A}if(S){if(bn(S)){const ae=await S(d,t),B=ap(ae,y);if(B&&(A[p]={...B,...G(ln.validate,B.message)},!n))return _(B.message),A}else if(Ke(S)){let ae={};for(const B in S){if(!Lt(ae)&&!n)break;const ie=ap(await S[B](d,t),y,B);ie&&(ae={...ie,...G(B,ie.message)},_(ie.message),n&&(A[p]=ae))}if(!Lt(ae)&&(A[p]={ref:y,...ae},!n))return A}}return _(!0),A};function OA(e,t){const n=t.slice(0,-1).length;let r=0;for(;r<n;)e=je(e)?r++:e[t[r++]];return e}function bA(e){for(const t in e)if(e.hasOwnProperty(t)&&!je(e[t]))return!1;return!0}function Xe(e,t){const n=Array.isArray(t)?t:Qf(t)?[t]:Hg(t),r=n.length===1?e:OA(e,n),o=n.length-1,i=n[o];return r&&delete r[i],o!==0&&(Ke(r)&&Lt(r)||Array.isArray(r)&&bA(r))&&Xe(e,n.slice(0,-1)),e}function qa(){let e=[];return{get observers(){return e},next:o=>{for(const i of e)i.next&&i.next(o)},subscribe:o=>(e.push(o),{unsubscribe:()=>{e=e.filter(i=>i!==o)}}),unsubscribe:()=>{e=[]}}}var dl=e=>at(e)||!Ug(e);function or(e,t){if(dl(e)||dl(t))return e===t;if(Ur(e)&&Ur(t))return e.getTime()===t.getTime();const n=Object.keys(e),r=Object.keys(t);if(n.length!==r.length)return!1;for(const o of n){const i=e[o];if(!r.includes(o))return!1;if(o!=="ref"){const s=t[o];if(Ur(i)&&Ur(s)||Ke(i)&&Ke(s)||Array.isArray(i)&&Array.isArray(s)?!or(i,s):i!==s)return!1}}return!0}var qg=e=>e.type==="select-multiple",FA=e=>Kf(e)||Di(e),Ka=e=>cl(e)&&e.isConnected,Kg=e=>{for(const t in e)if(bn(e[t]))return!0;return!1};function hl(e,t={}){const n=Array.isArray(e);if(Ke(e)||n)for(const r in e)Array.isArray(e[r])||Ke(e[r])&&!Kg(e[r])?(t[r]=Array.isArray(e[r])?[]:{},hl(e[r],t[r])):at(e[r])||(t[r]=!0);return t}function Gg(e,t,n){const r=Array.isArray(e);if(Ke(e)||r)for(const o in e)Array.isArray(e[o])||Ke(e[o])&&!Kg(e[o])?je(t)||dl(n[o])?n[o]=Array.isArray(e[o])?hl(e[o],[]):{...hl(e[o])}:Gg(e[o],at(t)?{}:t[o],n[o]):n[o]=!or(e[o],t[o]);return n}var Ga=(e,t)=>Gg(e,t,hl(t)),Yg=(e,{valueAsNumber:t,valueAsDate:n,setValueAs:r})=>je(e)?e:t?e===""?NaN:e&&+e:n&&en(e)?new Date(e):r?r(e):e;function Ya(e){const t=e.ref;if(!(e.refs?e.refs.every(n=>n.disabled):t.disabled))return qf(t)?t.files:Kf(t)?Qg(e.refs).value:qg(t)?[...t.selectedOptions].map(({value:n})=>n):Di(t)?Wg(e.refs).value:Yg(je(t.value)?e.ref.value:t.value,e)}var DA=(e,t,n,r)=>{const o={};for(const i of e){const s=W(t,i);s&&ke(o,i,s._f)}return{criteriaMode:n,names:[...e],fields:o,shouldUseNativeValidation:r}},bo=e=>je(e)?e:fl(e)?e.source:Ke(e)?fl(e.value)?e.value.source:e.value:e,MA=e=>e.mount&&(e.required||e.min||e.max||e.maxLength||e.minLength||e.pattern||e.validate);function cp(e,t,n){const r=W(e,n);if(r||Qf(n))return{error:r,name:n};const o=n.split(".");for(;o.length;){const i=o.join("."),s=W(t,i),l=W(e,i);if(s&&!Array.isArray(s)&&n!==i)return{name:n};if(l&&l.type)return{name:i,error:l};o.pop()}return{name:n}}var UA=(e,t,n,r,o)=>o.isOnAll?!1:!n&&o.isOnTouch?!(t||e):(n?r.isOnBlur:o.isOnBlur)?!e:(n?r.isOnChange:o.isOnChange)?e:!0,IA=(e,t)=>!Mi(W(e,t)).length&&Xe(e,t);const VA={mode:Wt.onSubmit,reValidateMode:Wt.onChange,shouldFocusError:!0};function $A(e={},t){let n={...VA,...e},r={submitCount:0,isDirty:!1,isLoading:bn(n.defaultValues),isValidating:!1,isSubmitted:!1,isSubmitting:!1,isSubmitSuccessful:!1,isValid:!1,touchedFields:{},dirtyFields:{},errors:{}},o={},i=Ke(n.defaultValues)||Ke(n.values)?Bt(n.defaultValues||n.values)||{}:{},s=n.shouldUnregister?{}:Bt(i),l={action:!1,mount:!1,watch:!1},a={mount:new Set,unMount:new Set,array:new Set,watch:new Set},u,f=0;const c={isDirty:!1,dirtyFields:!1,touchedFields:!1,isValidating:!1,isValid:!1,errors:!1},h={values:qa(),array:qa(),state:qa()},S=e.resetOptions&&e.resetOptions.keepDirtyValues,p=rp(n.mode),v=rp(n.reValidateMode),T=n.criteriaMode===Wt.all,m=g=>N=>{clearTimeout(f),f=setTimeout(g,N)},d=async g=>{if(c.isValid||g){const N=n.resolver?Lt((await J()).errors):await ye(o,!0);N!==r.isValid&&h.state.next({isValid:N})}},y=g=>c.isValidating&&h.state.next({isValidating:g}),_=(g,N=[],L,K,$=!0,I=!0)=>{if(K&&L){if(l.action=!0,I&&Array.isArray(W(o,g))){const ee=L(W(o,g),K.argA,K.argB);$&&ke(o,g,ee)}if(I&&Array.isArray(W(r.errors,g))){const ee=L(W(r.errors,g),K.argA,K.argB);$&&ke(r.errors,g,ee),IA(r.errors,g)}if(c.touchedFields&&I&&Array.isArray(W(r.touchedFields,g))){const ee=L(W(r.touchedFields,g),K.argA,K.argB);$&&ke(r.touchedFields,g,ee)}c.dirtyFields&&(r.dirtyFields=Ga(i,s)),h.state.next({name:g,isDirty:B(g,N),dirtyFields:r.dirtyFields,errors:r.errors,isValid:r.isValid})}else ke(s,g,N)},A=(g,N)=>{ke(r.errors,g,N),h.state.next({errors:r.errors})},k=(g,N,L,K)=>{const $=W(o,g);if($){const I=W(s,g,je(L)?W(i,g):L);je(I)||K&&K.defaultChecked||N?ke(s,g,N?I:Ya($._f)):ge(g,I),l.mount&&d()}},x=(g,N,L,K,$)=>{let I=!1,ee=!1;const Te={name:g};if(!L||K){c.isDirty&&(ee=r.isDirty,r.isDirty=Te.isDirty=B(),I=ee!==Te.isDirty);const Ae=or(W(i,g),N);ee=W(r.dirtyFields,g),Ae?Xe(r.dirtyFields,g):ke(r.dirtyFields,g,!0),Te.dirtyFields=r.dirtyFields,I=I||c.dirtyFields&&ee!==!Ae}if(L){const Ae=W(r.touchedFields,g);Ae||(ke(r.touchedFields,g,L),Te.touchedFields=r.touchedFields,I=I||c.touchedFields&&Ae!==L)}return I&&$&&h.state.next(Te),I?Te:{}},O=(g,N,L,K)=>{const $=W(r.errors,g),I=c.isValid&&On(N)&&r.isValid!==N;if(e.delayError&&L?(u=m(()=>A(g,L)),u(e.delayError)):(clearTimeout(f),u=null,L?ke(r.errors,g,L):Xe(r.errors,g)),(L?!or($,L):$)||!Lt(K)||I){const ee={...K,...I&&On(N)?{isValid:N}:{},errors:r.errors,name:g};r={...r,...ee},h.state.next(ee)}y(!1)},J=async g=>n.resolver(s,n.context,DA(g||a.mount,o,n.criteriaMode,n.shouldUseNativeValidation)),G=async g=>{const{errors:N}=await J(g);if(g)for(const L of g){const K=W(N,L);K?ke(r.errors,L,K):Xe(r.errors,L)}else r.errors=N;return N},ye=async(g,N,L={valid:!0})=>{for(const K in g){const $=g[K];if($){const{_f:I,...ee}=$;if(I){const Te=a.array.has(I.name),Ae=await up($,s,T,n.shouldUseNativeValidation&&!N,Te);if(Ae[I.name]&&(L.valid=!1,N))break;!N&&(W(Ae,I.name)?Te?PA(r.errors,Ae,I.name):ke(r.errors,I.name,Ae[I.name]):Xe(r.errors,I.name))}ee&&await ye(ee,N,L)}}return L.valid},ae=()=>{for(const g of a.unMount){const N=W(o,g);N&&(N._f.refs?N._f.refs.every(L=>!Ka(L)):!Ka(N._f.ref))&&j(g)}a.unMount=new Set},B=(g,N)=>(g&&N&&ke(s,g,N),!or(Se(),i)),ie=(g,N,L)=>zg(g,a,{...l.mount?s:je(N)?i:en(g)?{[g]:N}:N},L,N),pe=g=>Mi(W(l.mount?s:i,g,e.shouldUnregister?W(i,g,[]):[])),ge=(g,N,L={})=>{const K=W(o,g);let $=N;if(K){const I=K._f;I&&(!I.disabled&&ke(s,g,Yg(N,I)),$=cl(I.ref)&&at(N)?"":N,qg(I.ref)?[...I.ref.options].forEach(ee=>ee.selected=$.includes(ee.value)):I.refs?Di(I.ref)?I.refs.length>1?I.refs.forEach(ee=>(!ee.defaultChecked||!ee.disabled)&&(ee.checked=Array.isArray($)?!!$.find(Te=>Te===ee.value):$===ee.value)):I.refs[0]&&(I.refs[0].checked=!!$):I.refs.forEach(ee=>ee.checked=ee.value===$):qf(I.ref)?I.ref.value="":(I.ref.value=$,I.ref.type||h.values.next({name:g,values:{...s}})))}(L.shouldDirty||L.shouldTouch)&&x(g,$,L.shouldTouch,L.shouldDirty,!0),L.shouldValidate&&te(g)},_e=(g,N,L)=>{for(const K in N){const $=N[K],I=`${g}.${K}`,ee=W(o,I);(a.array.has(g)||!dl($)||ee&&!ee._f)&&!Ur($)?_e(I,$,L):ge(I,$,L)}},V=(g,N,L={})=>{const K=W(o,g),$=a.array.has(g),I=Bt(N);ke(s,g,I),$?(h.array.next({name:g,values:{...s}}),(c.isDirty||c.dirtyFields)&&L.shouldDirty&&h.state.next({name:g,dirtyFields:Ga(i,s),isDirty:B(g,I)})):K&&!K._f&&!at(I)?_e(g,I,L):ge(g,I,L),op(g,a)&&h.state.next({...r}),h.values.next({name:g,values:{...s}}),!l.mount&&t()},X=async g=>{const N=g.target;let L=N.name,K=!0;const $=W(o,L),I=()=>N.type?Ya($._f):Ig(g);if($){let ee,Te;const Ae=I(),wn=g.type===ul.BLUR||g.type===ul.FOCUS_OUT,sa=!MA($._f)&&!n.resolver&&!W(r.errors,L)&&!$._f.deps||UA(wn,W(r.touchedFields,L),r.isSubmitted,v,p),mo=op(L,a,wn);ke(s,L,Ae),wn?($._f.onBlur&&$._f.onBlur(g),u&&u(0)):$._f.onChange&&$._f.onChange(g);const yo=x(L,Ae,wn,!1),la=!Lt(yo)||mo;if(!wn&&h.values.next({name:L,type:g.type,values:{...s}}),sa)return c.isValid&&d(),la&&h.state.next({name:L,...mo?{}:yo});if(!wn&&mo&&h.state.next({...r}),y(!0),n.resolver){const{errors:Ui}=await J([L]),Ii=cp(r.errors,o,L),Vi=cp(Ui,o,Ii.name||L);ee=Vi.error,L=Vi.name,Te=Lt(Ui)}else ee=(await up($,s,T,n.shouldUseNativeValidation))[L],K=Number.isNaN(Ae)||Ae===W(s,L,Ae),K&&(ee?Te=!1:c.isValid&&(Te=await ye(o,!0)));K&&($._f.deps&&te($._f.deps),O(L,Te,ee,yo))}},te=async(g,N={})=>{let L,K;const $=Ts(g);if(y(!0),n.resolver){const I=await G(je(g)?g:$);L=Lt(I),K=g?!$.some(ee=>W(I,ee)):L}else g?(K=(await Promise.all($.map(async I=>{const ee=W(o,I);return await ye(ee&&ee._f?{[I]:ee}:ee)}))).every(Boolean),!(!K&&!r.isValid)&&d()):K=L=await ye(o);return h.state.next({...!en(g)||c.isValid&&L!==r.isValid?{}:{name:g},...n.resolver||!g?{isValid:L}:{},errors:r.errors,isValidating:!1}),N.shouldFocus&&!K&&pc(o,I=>I&&W(r.errors,I),g?$:a.mount),K},Se=g=>{const N={...i,...l.mount?s:{}};return je(g)?N:en(g)?W(N,g):g.map(L=>W(N,L))},E=(g,N)=>({invalid:!!W((N||r).errors,g),isDirty:!!W((N||r).dirtyFields,g),isTouched:!!W((N||r).touchedFields,g),error:W((N||r).errors,g)}),b=g=>{g&&Ts(g).forEach(N=>Xe(r.errors,N)),h.state.next({errors:g?r.errors:{}})},F=(g,N,L)=>{const K=(W(o,g,{_f:{}})._f||{}).ref;ke(r.errors,g,{...N,ref:K}),h.state.next({name:g,errors:r.errors,isValid:!1}),L&&L.shouldFocus&&K&&K.focus&&K.focus()},Y=(g,N)=>bn(g)?h.values.subscribe({next:L=>g(ie(void 0,N),L)}):ie(g,N,!0),j=(g,N={})=>{for(const L of g?Ts(g):a.mount)a.mount.delete(L),a.array.delete(L),N.keepValue||(Xe(o,L),Xe(s,L)),!N.keepError&&Xe(r.errors,L),!N.keepDirty&&Xe(r.dirtyFields,L),!N.keepTouched&&Xe(r.touchedFields,L),!n.shouldUnregister&&!N.keepDefaultValue&&Xe(i,L);h.values.next({values:{...s}}),h.state.next({...r,...N.keepDirty?{isDirty:B()}:{}}),!N.keepIsValid&&d()},le=({disabled:g,name:N,field:L,fields:K})=>{if(On(g)){const $=g?void 0:W(s,N,Ya(L?L._f:W(K,N)._f));ke(s,N,$),x(N,$,!1,!1,!0)}},ue=(g,N={})=>{let L=W(o,g);const K=On(N.disabled);return ke(o,g,{...L||{},_f:{...L&&L._f?L._f:{ref:{name:g}},name:g,mount:!0,...N}}),a.mount.add(g),L?le({field:L,disabled:N.disabled,name:g}):k(g,!0,N.value),{...K?{disabled:N.disabled}:{},...n.progressive?{required:!!N.required,min:bo(N.min),max:bo(N.max),minLength:bo(N.minLength),maxLength:bo(N.maxLength),pattern:bo(N.pattern)}:{},name:g,onChange:X,onBlur:X,ref:$=>{if($){ue(g,N),L=W(o,g);const I=je($.value)&&$.querySelectorAll&&$.querySelectorAll("input,select,textarea")[0]||$,ee=FA(I),Te=L._f.refs||[];if(ee?Te.find(Ae=>Ae===I):I===L._f.ref)return;ke(o,g,{_f:{...L._f,...ee?{refs:[...Te.filter(Ka),I,...Array.isArray(W(i,g))?[{}]:[]],ref:{type:I.type,name:g}}:{ref:I}}}),k(g,!1,void 0,I)}else L=W(o,g,{}),L._f&&(L._f.mount=!1),(n.shouldUnregister||N.shouldUnregister)&&!(Vg(a.array,g)&&l.action)&&a.unMount.add(g)}}},Z=()=>n.shouldFocusError&&pc(o,g=>g&&W(r.errors,g),a.mount),Ee=(g,N)=>async L=>{L&&(L.preventDefault&&L.preventDefault(),L.persist&&L.persist());let K=Bt(s);if(h.state.next({isSubmitting:!0}),n.resolver){const{errors:$,values:I}=await J();r.errors=$,K=I}else await ye(o);Xe(r.errors,"root"),Lt(r.errors)?(h.state.next({errors:{}}),await g(K,L)):(N&&await N({...r.errors},L),Z(),setTimeout(Z)),h.state.next({isSubmitted:!0,isSubmitting:!1,isSubmitSuccessful:Lt(r.errors),submitCount:r.submitCount+1,errors:r.errors})},ze=(g,N={})=>{W(o,g)&&(je(N.defaultValue)?V(g,W(i,g)):(V(g,N.defaultValue),ke(i,g,N.defaultValue)),N.keepTouched||Xe(r.touchedFields,g),N.keepDirty||(Xe(r.dirtyFields,g),r.isDirty=N.defaultValue?B(g,W(i,g)):B()),N.keepError||(Xe(r.errors,g),c.isValid&&d()),h.state.next({...r}))},we=(g,N={})=>{const L=g?Bt(g):i,K=Bt(L),$=g&&!Lt(g)?K:i;if(N.keepDefaultValues||(i=L),!N.keepValues){if(N.keepDirtyValues||S)for(const I of a.mount)W(r.dirtyFields,I)?ke($,I,W(s,I)):V(I,W($,I));else{if(zf&&je(g))for(const I of a.mount){const ee=W(o,I);if(ee&&ee._f){const Te=Array.isArray(ee._f.refs)?ee._f.refs[0]:ee._f.ref;if(cl(Te)){const Ae=Te.closest("form");if(Ae){Ae.reset();break}}}}o={}}s=e.shouldUnregister?N.keepDefaultValues?Bt(i):{}:Bt($),h.array.next({values:{...$}}),h.values.next({values:{...$}})}a={mount:new Set,unMount:new Set,array:new Set,watch:new Set,watchAll:!1,focus:""},!l.mount&&t(),l.mount=!c.isValid||!!N.keepIsValid,l.watch=!!e.shouldUnregister,h.state.next({submitCount:N.keepSubmitCount?r.submitCount:0,isDirty:N.keepDirty?r.isDirty:!!(N.keepDefaultValues&&!or(g,i)),isSubmitted:N.keepIsSubmitted?r.isSubmitted:!1,dirtyFields:N.keepDirtyValues?r.dirtyFields:N.keepDefaultValues&&g?Ga(i,g):{},touchedFields:N.keepTouched?r.touchedFields:{},errors:N.keepErrors?r.errors:{},isSubmitSuccessful:N.keepIsSubmitSuccessful?r.isSubmitSuccessful:!1,isSubmitting:!1})},Ce=(g,N)=>we(bn(g)?g(s):g,N);return{control:{register:ue,unregister:j,getFieldState:E,handleSubmit:Ee,setError:F,_executeSchema:J,_getWatch:ie,_getDirty:B,_updateValid:d,_removeUnmounted:ae,_updateFieldArray:_,_updateDisabledField:le,_getFieldArray:pe,_reset:we,_resetDefaultValues:()=>bn(n.defaultValues)&&n.defaultValues().then(g=>{Ce(g,n.resetOptions),h.state.next({isLoading:!1})}),_updateFormState:g=>{r={...r,...g}},_subjects:h,_proxyFormState:c,get _fields(){return o},get _formValues(){return s},get _state(){return l},set _state(g){l=g},get _defaultValues(){return i},get _names(){return a},set _names(g){a=g},get _formState(){return r},set _formState(g){r=g},get _options(){return n},set _options(g){n={...n,...g}}},trigger:te,register:ue,handleSubmit:Ee,watch:Y,setValue:V,getValues:Se,reset:Ce,resetField:ze,clearErrors:b,unregister:j,setError:F,setFocus:(g,N={})=>{const L=W(o,g),K=L&&L._f;if(K){const $=K.refs?K.refs[0]:K.ref;$.focus&&($.focus(),N.shouldSelect&&$.select())}},getFieldState:E}}function jA(e={}){const t=ne.useRef(),n=ne.useRef(),[r,o]=ne.useState({isDirty:!1,isValidating:!1,isLoading:bn(e.defaultValues),isSubmitted:!1,isSubmitting:!1,isSubmitSuccessful:!1,isValid:!1,submitCount:0,dirtyFields:{},touchedFields:{},errors:{},defaultValues:bn(e.defaultValues)?void 0:e.defaultValues});t.current||(t.current={...$A(e,()=>o(s=>({...s}))),formState:r});const i=t.current.control;return i._options=e,Wf({subject:i._subjects.state,next:s=>{jg(s,i._proxyFormState,i._updateFormState,!0)&&o({...i._formState})}}),ne.useEffect(()=>{e.values&&!or(e.values,n.current)?(i._reset(e.values,i._options.resetOptions),n.current=e.values):i._resetDefaultValues()},[e.values,i]),ne.useEffect(()=>{i._state.mount||(i._updateValid(),i._state.mount=!0),i._state.watch&&(i._state.watch=!1,i._subjects.state.next({...i._formState})),i._removeUnmounted()}),t.current.formState=$g(r,i),t.current}const Jg=({id:e,onCancel:t,onSubmit:n})=>{var _,A;const r=ea(),[o,i]=U.useState(!1),[,s]=Hn(Fi),[l,a]=U.useState(""),{data:u,error:f}=Ff({queryKey:["app-form",e],queryFn:()=>Zo.get(`/server/${e}`).then(k=>k.data),enabled:!!e}),{control:c,handleSubmit:h,reset:S,formState:{errors:p}}=jA({defaultValues:{display_name:"",description:"",framework:"",thumbnail:"",filepath:"",conda_env:"",custom_command:"",profile:""}}),v=({display_name:k,description:x,framework:O,thumbnail:J,filepath:G,conda_env:ye,custom_command:ae,profile:B})=>{const ie={servername:l||k,user_options:{jhub_app:!0,name:l||k,display_name:k,description:x||"",framework:O,thumbnail:J||"",filepath:G||"",conda_env:ye||"",custom_command:ae||"",profile:B||""}};i(!0),e?y(ie,{onSuccess:async()=>{i(!1),r.invalidateQueries(["app-state"]),n&&n()},onError:async pe=>{i(!1),s(pe.message)}}):d(ie,{onSuccess:async pe=>{i(!1);const ge=Bf().user;if(ge&&(pe==null?void 0:pe.length)>1){const _e=pe[1];window.location.assign(`/user/${ge}/${_e}`)}},onError:async pe=>{i(!1),s(pe.message)}})},T=async({servername:k,user_options:x})=>(await Zo.post("/server",{servername:k,user_options:x})).data,m=async({servername:k,user_options:x})=>(await Zo.put(`/server/${k}`,{servername:k,user_options:x})).data,{mutate:d}=lc({mutationFn:T,retry:1}),{mutate:y}=lc({mutationFn:m,retry:1});return U.useEffect(()=>{u!=null&&u.name&&(u!=null&&u.user_options)&&(a(u.name),S({...u.user_options}))},[u==null?void 0:u.name,u==null?void 0:u.user_options,S]),U.useEffect(()=>{f&&s(f.message)},[f,s]),C.jsxs("form",{id:"app-form",onSubmit:h(v),className:"form",children:[C.jsxs(Er,{children:[C.jsx(Cr,{htmlFor:"display_name",required:!0,children:"Display Name"}),C.jsx(_r,{name:"display_name",control:c,rules:np,render:({field:{ref:k,...x}})=>C.jsx($o,{...x,id:"display_name",autoFocus:!0})}),((_=p.display_name)==null?void 0:_.message)&&C.jsx(fp,{errors:[p.display_name.message]})]}),C.jsxs(Er,{children:[C.jsx(Cr,{htmlFor:"description",children:"Description"}),C.jsx(_r,{name:"description",control:c,render:({field:{ref:k,...x}})=>C.jsx(zA,{...x,id:"description"})})]}),C.jsxs(Er,{children:[C.jsx(Cr,{htmlFor:"framework",required:!0,children:"Framework"}),C.jsx(_r,{name:"framework",control:c,rules:np,render:({field:{ref:k,...x}})=>C.jsx(BA,{...x,id:"framework",options:[{value:"",label:"Select..."},{value:"panel",label:"Panel"},{value:"bokeh",label:"Bokeh"},{value:"streamlit",label:"Streamlit"},{value:"voila",label:"Voila"},{value:"plotlydash",label:"PlotlyDash"},{value:"gradio",label:"Gradio"},{value:"jupyterlab",label:"JupyterLab"},{value:"custom",label:"Custom Command"}]})}),((A=p.framework)==null?void 0:A.message)&&C.jsx(fp,{errors:[p.framework.message]})]}),C.jsxs(Er,{children:[C.jsx(Cr,{htmlFor:"filepath",children:"Filepath"}),C.jsx(_r,{name:"filepath",control:c,render:({field:{ref:k,...x}})=>C.jsx($o,{...x,id:"filepath"})})]}),C.jsxs(Er,{children:[C.jsx(Cr,{htmlFor:"conda_env",children:"Conda Environment"}),C.jsx(_r,{name:"conda_env",control:c,render:({field:{ref:k,...x}})=>C.jsx($o,{...x,id:"conda_env"})})]}),C.jsxs(Er,{children:[C.jsx(Cr,{htmlFor:"custom_command",children:"Custom Command"}),C.jsx(_r,{name:"custom_command",control:c,render:({field:{ref:k,...x}})=>C.jsx($o,{...x,id:"custom_command"})})]}),C.jsxs(Xg,{children:[C.jsx(so,{id:"cancel-btn",type:"button",variant:"secondary",onClick:t,children:"Cancel"}),C.jsx(so,{id:"submit-btn",type:"submit",disabled:o,children:"Submit"})]})]})},Xg=({id:e=void 0,className:t,children:n})=>{const r=Yn("button-group",t);return C.jsx("ul",{id:e,className:r,children:U.Children.map(n,(o,i)=>C.jsx("li",{className:"button-group-item",children:o},i))})},fp=({id:e=void 0,errors:t,children:n})=>!n&&!t?C.jsx(C.Fragment,{}):C.jsx(C.Fragment,{children:n??(t==null?void 0:t.map((r,o)=>C.jsx("span",{id:`${e}-${o}`,className:"text-red",children:r},o)))}),Er=({id:e=void 0,errors:t,className:n,children:r})=>{const o=!!(t&&t.length>0),i=Yn("form-group",{"form-group-error":o},n);return C.jsx("div",{id:e,className:i,children:r})},Cr=({htmlFor:e,required:t,children:n,...r})=>C.jsxs("label",{className:"label",htmlFor:e,...r,children:[n,t&&C.jsx("span",{className:"text-red",children:" *"})]}),BA=({id:e,options:t,className:n,onChange:r,...o})=>{const i=Yn("select p-2",n);return C.jsx("select",{id:e,className:i,onChange:r,...o,children:t.map((s,l)=>C.jsx("option",{value:s.value,children:s.label},l))})},zA=({id:e,rows:t,className:n,onChange:r,...o})=>C.jsx("textarea",{className:Yn("text-area p-2",n),id:e,rows:t,onChange:r,...o}),$o=({id:e,name:t,className:n,type:r,onChange:o,...i})=>{const s=Yn("text-input p-2",n);return C.jsx("input",{id:e,name:t,className:s,type:r,onChange:o,...i})},dp=({appType:e="My",filter:t})=>{const[n]=Hn(jf),[,r]=Hn(Fi),[o,i]=U.useState([]),{isLoading:s,error:l,data:a}=Ff({queryKey:["app-state"],queryFn:()=>Zo.get("/server/").then(u=>u.data).then(u=>u),enabled:!!n.user});return U.useEffect(()=>{if(!s&&a){const u=t.toLowerCase();i(()=>RA(a,e).filter(f=>{var c,h;return f.name.toLowerCase().includes(u)||((c=f.description)==null?void 0:c.toLowerCase().includes(u))||((h=f.framework)==null?void 0:h.toLowerCase().includes(u))}))}},[s,a,e,t]),U.useEffect(()=>{r(l?l.message:void 0)},[l,r]),C.jsxs(C.Fragment,{children:[C.jsxs("div",{className:"container grid grid-cols-12 flex flex-align-center pb-12",children:[C.jsx("div",{className:"col-span-1",children:C.jsxs("h2",{className:"whitespace-nowrap font-bold",children:[e," Apps"]})}),C.jsx("div",{className:"col-span-10",children:C.jsx("hr",{className:"spacer"})}),C.jsx("div",{className:"col-span-1 flex justify-end",children:C.jsxs("h2",{className:"whitespace-nowrap font-bold",children:[o.length," apps"]})})]}),C.jsx("div",{className:"container grid pb-12",children:o.length>0?C.jsx("div",{className:"flex flex flex-row flex-wrap gap-4",children:o.map(u=>C.jsx(SA,{id:u.id,title:u.name,description:u.description,thumbnail:u.thumbnail,framework:u.framework,url:u.url,ready:u.ready},`app-${u.id}`))}):C.jsx("div",{children:"No apps available"})})]})},HA=Mg.create({baseURL:"/api",headers:{"Content-Type":"application/json",Authorization:"Authorization: token 13748105b2e64b9082e10f802b04f523"},params:{_xsrf:Bf().xsrf_token}}),WA=()=>{const[e]=Hn(jf),[,t]=Hn(Fi),[n,r]=U.useState([]),{isLoading:o,error:i,data:s}=Ff({queryKey:["service-data"],queryFn:()=>HA.get("/services").then(a=>a.data).then(a=>a),enabled:!!e.user}),l=(a,u)=>{u?window.open(a,"_blank"):window.location.assign(a)};return U.useEffect(()=>{!o&&s&&r(()=>_A(s,e.user))},[o,s,e.user]),U.useEffect(()=>{t(i?i.message:void 0)},[i,t]),C.jsxs(C.Fragment,{children:[C.jsxs("div",{className:"container grid grid-cols-12 flex flex-align-center pb-12",children:[C.jsx("div",{className:"col-span-1",children:C.jsx("h2",{className:"whitespace-nowrap font-bold",children:"Services"})}),C.jsx("div",{className:"col-span-10",children:C.jsx("hr",{className:"spacer"})}),C.jsx("div",{className:"col-span-1 flex justify-end",children:C.jsxs("h2",{className:"whitespace-nowrap font-bold",children:[n.length," services"]})})]}),C.jsx("div",{className:"container grid pb-12",children:n.length>0?C.jsx("div",{className:"flex flex flex-row flex-wrap gap-4",children:n.map((a,u)=>C.jsx(so,{id:`service-${u}`,variant:"secondary",style:{minWidth:"180px"},onClick:()=>{l(a.url,a.external)},children:a.name},`service-${u}`))}):C.jsx("div",{children:"No services available"})})]})},hp=()=>{const[e,t]=U.useState(!1),[n]=Hn(Fi),[r,o]=U.useState(""),i=s=>{const l=s.target;o(l.value)};return C.jsxs(C.Fragment,{children:[C.jsxs("div",{className:"container grid grid-cols-12 pb-12",children:[C.jsx("div",{className:"md:col-span-2 xs:col-span-12",children:C.jsx("h1",{className:"text-3xl font-bold",children:"Home"})}),C.jsx("div",{className:"md:col-span-8 xs:col-span-8",children:C.jsx($o,{id:"search",placeholder:"Search...","aria-label":"Search for an app",className:"w-full mt-0",onChange:i})}),C.jsxs("div",{className:"md:col-span-2 xs:col-span-4 flex justify-end",children:[C.jsx(so,{id:"create-app",onClick:()=>{t(!0)},children:"Create App"}),e&&C.jsx(hc,{title:"Create New App",setIsOpen:t,body:C.jsx(Jg,{onCancel:()=>t(!1)})})]})]}),n&&C.jsx("div",{className:"container grid grid-cols-12 pb-2",children:C.jsx("div",{className:"col-span-12",children:C.jsx(qk,{id:"alert-notification",type:"error",children:n})})}),C.jsx(WA,{}),C.jsx(dp,{appType:"My",filter:r}),C.jsx(dp,{appType:"Shared",filter:r})]})},QA=new vk,qA=()=>{const[,e]=Hn(jf);return U.useEffect(()=>{e(Bf())},[e]),C.jsx(Ok,{client:QA,children:C.jsx("div",{children:C.jsx("main",{className:"my-6",children:C.jsxs(u_,{children:[C.jsx(Qu,{path:"/home",element:C.jsx(hp,{})}),C.jsx(Qu,{path:"/",element:C.jsx(hp,{})})]})})})})};Ja.createRoot(document.getElementById("root")).render(C.jsx(ne.StrictMode,{children:C.jsx(f_,{basename:"/hub",children:C.jsx(nk,{children:C.jsx(qA,{})})})}));
+*/(function(e){(function(){var t={}.hasOwnProperty;function n(){for(var r=[],o=0;o<arguments.length;o++){var i=arguments[o];if(i){var s=typeof i;if(s==="string"||s==="number")r.push(i);else if(Array.isArray(i)){if(i.length){var l=n.apply(null,i);l&&r.push(l)}}else if(s==="object"){if(i.toString!==Object.prototype.toString&&!i.toString.toString().includes("[native code]")){r.push(i.toString());continue}for(var a in i)t.call(i,a)&&i[a]&&r.push(a)}}}return r.join(" ")}e.exports?(n.default=n,e.exports=n):window.classNames=n})()})(pg);var Qk=pg.exports;const Yn=mc(Qk),qk=({id:e,type:t="info",children:n})=>{const r=Yn("alert",{"alert-success":t==="success","alert-warning":t==="warning","alert-error":t==="error","alert-info":t==="info"});return C.jsx("div",{id:e,className:r,children:C.jsx("div",{className:"alert-body",children:C.jsx("p",{className:"alert-text",children:n})})})},mg=({id:e=void 0,className:t,children:n})=>{const r=Yn("button-group",t);return C.jsx("ul",{id:e,className:r,children:U.Children.map(n,(o,i)=>C.jsx("li",{className:"button-group-item",children:o},i))})},io=({id:e,type:t="button",variant:n="primary",className:r,children:o,...i})=>{const s=Yn("btn rounded border-2 font-bold whitespace-nowrap",{"btn-primary":n==="primary","btn-secondary":n==="secondary"},r);return C.jsx("button",{id:e,type:t,className:s,...i,children:o})};function Kk({title:e,titleId:t,...n},r){return U.createElement("svg",Object.assign({xmlns:"http://www.w3.org/2000/svg",viewBox:"0 0 24 24",fill:"currentColor","aria-hidden":"true",ref:r,"aria-labelledby":t},n),e?U.createElement("title",{id:t},e):null,U.createElement("path",{fillRule:"evenodd",d:"M4.5 12a1.5 1.5 0 113 0 1.5 1.5 0 01-3 0zm6 0a1.5 1.5 0 113 0 1.5 1.5 0 01-3 0zm6 0a1.5 1.5 0 113 0 1.5 1.5 0 01-3 0z",clipRule:"evenodd"}))}const Gk=U.forwardRef(Kk),Yk=Gk;function Jk({title:e,titleId:t,...n},r){return U.createElement("svg",Object.assign({xmlns:"http://www.w3.org/2000/svg",viewBox:"0 0 24 24",fill:"currentColor","aria-hidden":"true",ref:r,"aria-labelledby":t},n),e?U.createElement("title",{id:t},e):null,U.createElement("path",{fillRule:"evenodd",d:"M5.47 5.47a.75.75 0 011.06 0L12 10.94l5.47-5.47a.75.75 0 111.06 1.06L13.06 12l5.47 5.47a.75.75 0 11-1.06 1.06L12 13.06l-5.47 5.47a.75.75 0 01-1.06-1.06L10.94 12 5.47 6.53a.75.75 0 010-1.06z",clipRule:"evenodd"}))}const Xk=U.forwardRef(Jk),Zk=Xk,eN=({id:e,items:t})=>{const n=ne.useRef(null);return C.jsxs("div",{className:"context-menu-container",id:e,tabIndex:0,ref:n,children:[C.jsx(Yk,{}),C.jsx("ul",{className:"context-menu",children:t.filter(r=>r.visible).map(r=>C.jsx("li",{children:C.jsx("a",{className:r.disabled?"disabled":"",onClick:o=>{var i;r.onClick&&!r.disabled&&(r.onClick(o),(i=n.current)==null||i.blur())},children:r.title})},`context-menu-item-${r.id}`))})]})},Wh=({id:e=void 0,errors:t,children:n})=>!n&&!t?C.jsx(C.Fragment,{}):C.jsx(C.Fragment,{children:n??(t==null?void 0:t.map((r,o)=>C.jsx("span",{id:`${e}-${o}`,className:"text-red",children:r},o)))}),_r=({id:e=void 0,errors:t,className:n,children:r})=>{const o=!!(t&&t.length>0),i=Yn("form-group",{"form-group-error":o},n);return C.jsx("div",{id:e,className:i,children:r})},Rr=({htmlFor:e,required:t,children:n,...r})=>C.jsxs("label",{className:"label",htmlFor:e,...r,children:[n,t&&C.jsx("span",{className:"text-red",children:" *"})]}),ac=({title:e,setIsOpen:t,body:n,footer:r})=>C.jsxs(C.Fragment,{children:[C.jsx("div",{className:"modal-overlay",onClick:()=>t(!1)}),C.jsx("div",{className:"modal-container",children:C.jsxs("div",{className:"modal-main",children:[C.jsxs("div",{className:"modal-heading",children:[C.jsx("h5",{className:"modal-title",children:e}),C.jsx("button",{className:"modal-close-btn",onClick:()=>t(!1),children:C.jsx(Zk,{})})]}),C.jsx("div",{className:"modal-body",children:n}),r?C.jsx("div",{className:"modal-footer",children:r}):C.jsx(C.Fragment,{})]})})]}),tN=({id:e,options:t,className:n,onChange:r,...o})=>{const i=Yn("select p-2",n);return C.jsx("select",{id:e,className:i,onChange:r,...o,children:t.map((s,l)=>C.jsx("option",{value:s.value,children:s.label},l))})},nN=({id:e,children:t,className:n})=>{const r=Yn("tag",n);return C.jsx("span",{id:e,className:r,children:t})},rN=({id:e,rows:t,className:n,onChange:r,...o})=>C.jsx("textarea",{className:Yn("text-area p-2",n),id:e,rows:t,onChange:r,...o}),$o=({id:e,name:t,className:n,type:r,onChange:o,...i})=>{const s=Yn("text-input p-2",n);return C.jsx("input",{id:e,name:t,className:s,type:r,onChange:o,...i})};Lf({key:"currentUser",default:void 0});const Df=Lf({key:"currentJhData",default:{admin_access:!1,base_url:"/hub",options_form:!1,prefix:"/",user:"",xsrf_token:""}}),Oi=Lf({key:"currentNotification",default:void 0});function yg(e,t){return function(){return e.apply(t,arguments)}}const{toString:oN}=Object.prototype,{getPrototypeOf:Mf}=Object,ta=(e=>t=>{const n=oN.call(t);return e[n]||(e[n]=n.slice(8,-1).toLowerCase())})(Object.create(null)),sn=e=>(e=e.toLowerCase(),t=>ta(t)===e),na=e=>t=>typeof t===e,{isArray:po}=Array,gi=na("undefined");function iN(e){return e!==null&&!gi(e)&&e.constructor!==null&&!gi(e.constructor)&&Dt(e.constructor.isBuffer)&&e.constructor.isBuffer(e)}const vg=sn("ArrayBuffer");function sN(e){let t;return typeof ArrayBuffer<"u"&&ArrayBuffer.isView?t=ArrayBuffer.isView(e):t=e&&e.buffer&&vg(e.buffer),t}const lN=na("string"),Dt=na("function"),gg=na("number"),ra=e=>e!==null&&typeof e=="object",aN=e=>e===!0||e===!1,Es=e=>{if(ta(e)!=="object")return!1;const t=Mf(e);return(t===null||t===Object.prototype||Object.getPrototypeOf(t)===null)&&!(Symbol.toStringTag in e)&&!(Symbol.iterator in e)},uN=sn("Date"),cN=sn("File"),fN=sn("Blob"),dN=sn("FileList"),hN=e=>ra(e)&&Dt(e.pipe),pN=e=>{let t;return e&&(typeof FormData=="function"&&e instanceof FormData||Dt(e.append)&&((t=ta(e))==="formdata"||t==="object"&&Dt(e.toString)&&e.toString()==="[object FormData]"))},mN=sn("URLSearchParams"),yN=e=>e.trim?e.trim():e.replace(/^[\s\uFEFF\xA0]+|[\s\uFEFF\xA0]+$/g,"");function bi(e,t,{allOwnKeys:n=!1}={}){if(e===null||typeof e>"u")return;let r,o;if(typeof e!="object"&&(e=[e]),po(e))for(r=0,o=e.length;r<o;r++)t.call(null,e[r],r,e);else{const i=n?Object.getOwnPropertyNames(e):Object.keys(e),s=i.length;let l;for(r=0;r<s;r++)l=i[r],t.call(null,e[l],l,e)}}function Sg(e,t){t=t.toLowerCase();const n=Object.keys(e);let r=n.length,o;for(;r-- >0;)if(o=n[r],t===o.toLowerCase())return o;return null}const wg=(()=>typeof globalThis<"u"?globalThis:typeof self<"u"?self:typeof window<"u"?window:global)(),_g=e=>!gi(e)&&e!==wg;function uc(){const{caseless:e}=_g(this)&&this||{},t={},n=(r,o)=>{const i=e&&Sg(t,o)||o;Es(t[i])&&Es(r)?t[i]=uc(t[i],r):Es(r)?t[i]=uc({},r):po(r)?t[i]=r.slice():t[i]=r};for(let r=0,o=arguments.length;r<o;r++)arguments[r]&&bi(arguments[r],n);return t}const vN=(e,t,n,{allOwnKeys:r}={})=>(bi(t,(o,i)=>{n&&Dt(o)?e[i]=yg(o,n):e[i]=o},{allOwnKeys:r}),e),gN=e=>(e.charCodeAt(0)===65279&&(e=e.slice(1)),e),SN=(e,t,n,r)=>{e.prototype=Object.create(t.prototype,r),e.prototype.constructor=e,Object.defineProperty(e,"super",{value:t.prototype}),n&&Object.assign(e.prototype,n)},wN=(e,t,n,r)=>{let o,i,s;const l={};if(t=t||{},e==null)return t;do{for(o=Object.getOwnPropertyNames(e),i=o.length;i-- >0;)s=o[i],(!r||r(s,e,t))&&!l[s]&&(t[s]=e[s],l[s]=!0);e=n!==!1&&Mf(e)}while(e&&(!n||n(e,t))&&e!==Object.prototype);return t},_N=(e,t,n)=>{e=String(e),(n===void 0||n>e.length)&&(n=e.length),n-=t.length;const r=e.indexOf(t,n);return r!==-1&&r===n},RN=e=>{if(!e)return null;if(po(e))return e;let t=e.length;if(!gg(t))return null;const n=new Array(t);for(;t-- >0;)n[t]=e[t];return n},EN=(e=>t=>e&&t instanceof e)(typeof Uint8Array<"u"&&Mf(Uint8Array)),CN=(e,t)=>{const r=(e&&e[Symbol.iterator]).call(e);let o;for(;(o=r.next())&&!o.done;){const i=o.value;t.call(e,i[0],i[1])}},xN=(e,t)=>{let n;const r=[];for(;(n=e.exec(t))!==null;)r.push(n);return r},TN=sn("HTMLFormElement"),kN=e=>e.toLowerCase().replace(/[-_\s]([a-z\d])(\w*)/g,function(n,r,o){return r.toUpperCase()+o}),Qh=(({hasOwnProperty:e})=>(t,n)=>e.call(t,n))(Object.prototype),NN=sn("RegExp"),Rg=(e,t)=>{const n=Object.getOwnPropertyDescriptors(e),r={};bi(n,(o,i)=>{let s;(s=t(o,i,e))!==!1&&(r[i]=s||o)}),Object.defineProperties(e,r)},AN=e=>{Rg(e,(t,n)=>{if(Dt(e)&&["arguments","caller","callee"].indexOf(n)!==-1)return!1;const r=e[n];if(Dt(r)){if(t.enumerable=!1,"writable"in t){t.writable=!1;return}t.set||(t.set=()=>{throw Error("Can not rewrite read-only method '"+n+"'")})}})},LN=(e,t)=>{const n={},r=o=>{o.forEach(i=>{n[i]=!0})};return po(e)?r(e):r(String(e).split(t)),n},PN=()=>{},ON=(e,t)=>(e=+e,Number.isFinite(e)?e:t),za="abcdefghijklmnopqrstuvwxyz",qh="0123456789",Eg={DIGIT:qh,ALPHA:za,ALPHA_DIGIT:za+za.toUpperCase()+qh},bN=(e=16,t=Eg.ALPHA_DIGIT)=>{let n="";const{length:r}=t;for(;e--;)n+=t[Math.random()*r|0];return n};function FN(e){return!!(e&&Dt(e.append)&&e[Symbol.toStringTag]==="FormData"&&e[Symbol.iterator])}const DN=e=>{const t=new Array(10),n=(r,o)=>{if(ra(r)){if(t.indexOf(r)>=0)return;if(!("toJSON"in r)){t[o]=r;const i=po(r)?[]:{};return bi(r,(s,l)=>{const a=n(s,o+1);!gi(a)&&(i[l]=a)}),t[o]=void 0,i}}return r};return n(e,0)},MN=sn("AsyncFunction"),UN=e=>e&&(ra(e)||Dt(e))&&Dt(e.then)&&Dt(e.catch),P={isArray:po,isArrayBuffer:vg,isBuffer:iN,isFormData:pN,isArrayBufferView:sN,isString:lN,isNumber:gg,isBoolean:aN,isObject:ra,isPlainObject:Es,isUndefined:gi,isDate:uN,isFile:cN,isBlob:fN,isRegExp:NN,isFunction:Dt,isStream:hN,isURLSearchParams:mN,isTypedArray:EN,isFileList:dN,forEach:bi,merge:uc,extend:vN,trim:yN,stripBOM:gN,inherits:SN,toFlatObject:wN,kindOf:ta,kindOfTest:sn,endsWith:_N,toArray:RN,forEachEntry:CN,matchAll:xN,isHTMLForm:TN,hasOwnProperty:Qh,hasOwnProp:Qh,reduceDescriptors:Rg,freezeMethods:AN,toObjectSet:LN,toCamelCase:kN,noop:PN,toFiniteNumber:ON,findKey:Sg,global:wg,isContextDefined:_g,ALPHABET:Eg,generateString:bN,isSpecCompliantForm:FN,toJSONObject:DN,isAsyncFn:MN,isThenable:UN};function me(e,t,n,r,o){Error.call(this),Error.captureStackTrace?Error.captureStackTrace(this,this.constructor):this.stack=new Error().stack,this.message=e,this.name="AxiosError",t&&(this.code=t),n&&(this.config=n),r&&(this.request=r),o&&(this.response=o)}P.inherits(me,Error,{toJSON:function(){return{message:this.message,name:this.name,description:this.description,number:this.number,fileName:this.fileName,lineNumber:this.lineNumber,columnNumber:this.columnNumber,stack:this.stack,config:P.toJSONObject(this.config),code:this.code,status:this.response&&this.response.status?this.response.status:null}}});const Cg=me.prototype,xg={};["ERR_BAD_OPTION_VALUE","ERR_BAD_OPTION","ECONNABORTED","ETIMEDOUT","ERR_NETWORK","ERR_FR_TOO_MANY_REDIRECTS","ERR_DEPRECATED","ERR_BAD_RESPONSE","ERR_BAD_REQUEST","ERR_CANCELED","ERR_NOT_SUPPORT","ERR_INVALID_URL"].forEach(e=>{xg[e]={value:e}});Object.defineProperties(me,xg);Object.defineProperty(Cg,"isAxiosError",{value:!0});me.from=(e,t,n,r,o,i)=>{const s=Object.create(Cg);return P.toFlatObject(e,s,function(a){return a!==Error.prototype},l=>l!=="isAxiosError"),me.call(s,e.message,t,n,r,o),s.cause=e,s.name=e.name,i&&Object.assign(s,i),s};const IN=null;function cc(e){return P.isPlainObject(e)||P.isArray(e)}function Tg(e){return P.endsWith(e,"[]")?e.slice(0,-2):e}function Kh(e,t,n){return e?e.concat(t).map(function(o,i){return o=Tg(o),!n&&i?"["+o+"]":o}).join(n?".":""):t}function VN(e){return P.isArray(e)&&!e.some(cc)}const $N=P.toFlatObject(P,{},null,function(t){return/^is[A-Z]/.test(t)});function oa(e,t,n){if(!P.isObject(e))throw new TypeError("target must be an object");t=t||new FormData,n=P.toFlatObject(n,{metaTokens:!0,dots:!1,indexes:!1},!1,function(v,T){return!P.isUndefined(T[v])});const r=n.metaTokens,o=n.visitor||f,i=n.dots,s=n.indexes,a=(n.Blob||typeof Blob<"u"&&Blob)&&P.isSpecCompliantForm(t);if(!P.isFunction(o))throw new TypeError("visitor must be a function");function u(p){if(p===null)return"";if(P.isDate(p))return p.toISOString();if(!a&&P.isBlob(p))throw new me("Blob is not supported. Use a Buffer instead.");return P.isArrayBuffer(p)||P.isTypedArray(p)?a&&typeof Blob=="function"?new Blob([p]):Buffer.from(p):p}function f(p,v,T){let m=p;if(p&&!T&&typeof p=="object"){if(P.endsWith(v,"{}"))v=r?v:v.slice(0,-2),p=JSON.stringify(p);else if(P.isArray(p)&&VN(p)||(P.isFileList(p)||P.endsWith(v,"[]"))&&(m=P.toArray(p)))return v=Tg(v),m.forEach(function(y,_){!(P.isUndefined(y)||y===null)&&t.append(s===!0?Kh([v],_,i):s===null?v:v+"[]",u(y))}),!1}return cc(p)?!0:(t.append(Kh(T,v,i),u(p)),!1)}const c=[],h=Object.assign($N,{defaultVisitor:f,convertValue:u,isVisitable:cc});function S(p,v){if(!P.isUndefined(p)){if(c.indexOf(p)!==-1)throw Error("Circular reference detected in "+v.join("."));c.push(p),P.forEach(p,function(m,d){(!(P.isUndefined(m)||m===null)&&o.call(t,m,P.isString(d)?d.trim():d,v,h))===!0&&S(m,v?v.concat(d):[d])}),c.pop()}}if(!P.isObject(e))throw new TypeError("data must be an object");return S(e),t}function Gh(e){const t={"!":"%21","'":"%27","(":"%28",")":"%29","~":"%7E","%20":"+","%00":"\0"};return encodeURIComponent(e).replace(/[!'()~]|%20|%00/g,function(r){return t[r]})}function Uf(e,t){this._pairs=[],e&&oa(e,this,t)}const kg=Uf.prototype;kg.append=function(t,n){this._pairs.push([t,n])};kg.toString=function(t){const n=t?function(r){return t.call(this,r,Gh)}:Gh;return this._pairs.map(function(o){return n(o[0])+"="+n(o[1])},"").join("&")};function jN(e){return encodeURIComponent(e).replace(/%3A/gi,":").replace(/%24/g,"$").replace(/%2C/gi,",").replace(/%20/g,"+").replace(/%5B/gi,"[").replace(/%5D/gi,"]")}function Ng(e,t,n){if(!t)return e;const r=n&&n.encode||jN,o=n&&n.serialize;let i;if(o?i=o(t,n):i=P.isURLSearchParams(t)?t.toString():new Uf(t,n).toString(r),i){const s=e.indexOf("#");s!==-1&&(e=e.slice(0,s)),e+=(e.indexOf("?")===-1?"?":"&")+i}return e}class BN{constructor(){this.handlers=[]}use(t,n,r){return this.handlers.push({fulfilled:t,rejected:n,synchronous:r?r.synchronous:!1,runWhen:r?r.runWhen:null}),this.handlers.length-1}eject(t){this.handlers[t]&&(this.handlers[t]=null)}clear(){this.handlers&&(this.handlers=[])}forEach(t){P.forEach(this.handlers,function(r){r!==null&&t(r)})}}const Yh=BN,Ag={silentJSONParsing:!0,forcedJSONParsing:!0,clarifyTimeoutError:!1},zN=typeof URLSearchParams<"u"?URLSearchParams:Uf,HN=typeof FormData<"u"?FormData:null,WN=typeof Blob<"u"?Blob:null,QN={isBrowser:!0,classes:{URLSearchParams:zN,FormData:HN,Blob:WN},protocols:["http","https","file","blob","url","data"]},Lg=typeof window<"u"&&typeof document<"u",qN=(e=>Lg&&["ReactNative","NativeScript","NS"].indexOf(e)<0)(typeof navigator<"u"&&navigator.product),KN=(()=>typeof WorkerGlobalScope<"u"&&self instanceof WorkerGlobalScope&&typeof self.importScripts=="function")(),GN=Object.freeze(Object.defineProperty({__proto__:null,hasBrowserEnv:Lg,hasStandardBrowserEnv:qN,hasStandardBrowserWebWorkerEnv:KN},Symbol.toStringTag,{value:"Module"})),Zt={...GN,...QN};function YN(e,t){return oa(e,new Zt.classes.URLSearchParams,Object.assign({visitor:function(n,r,o,i){return Zt.isNode&&P.isBuffer(n)?(this.append(r,n.toString("base64")),!1):i.defaultVisitor.apply(this,arguments)}},t))}function JN(e){return P.matchAll(/\w+|\[(\w*)]/g,e).map(t=>t[0]==="[]"?"":t[1]||t[0])}function XN(e){const t={},n=Object.keys(e);let r;const o=n.length;let i;for(r=0;r<o;r++)i=n[r],t[i]=e[i];return t}function Pg(e){function t(n,r,o,i){let s=n[i++];const l=Number.isFinite(+s),a=i>=n.length;return s=!s&&P.isArray(o)?o.length:s,a?(P.hasOwnProp(o,s)?o[s]=[o[s],r]:o[s]=r,!l):((!o[s]||!P.isObject(o[s]))&&(o[s]=[]),t(n,r,o[s],i)&&P.isArray(o[s])&&(o[s]=XN(o[s])),!l)}if(P.isFormData(e)&&P.isFunction(e.entries)){const n={};return P.forEachEntry(e,(r,o)=>{t(JN(r),o,n,0)}),n}return null}function ZN(e,t,n){if(P.isString(e))try{return(t||JSON.parse)(e),P.trim(e)}catch(r){if(r.name!=="SyntaxError")throw r}return(n||JSON.stringify)(e)}const If={transitional:Ag,adapter:["xhr","http"],transformRequest:[function(t,n){const r=n.getContentType()||"",o=r.indexOf("application/json")>-1,i=P.isObject(t);if(i&&P.isHTMLForm(t)&&(t=new FormData(t)),P.isFormData(t))return o&&o?JSON.stringify(Pg(t)):t;if(P.isArrayBuffer(t)||P.isBuffer(t)||P.isStream(t)||P.isFile(t)||P.isBlob(t))return t;if(P.isArrayBufferView(t))return t.buffer;if(P.isURLSearchParams(t))return n.setContentType("application/x-www-form-urlencoded;charset=utf-8",!1),t.toString();let l;if(i){if(r.indexOf("application/x-www-form-urlencoded")>-1)return YN(t,this.formSerializer).toString();if((l=P.isFileList(t))||r.indexOf("multipart/form-data")>-1){const a=this.env&&this.env.FormData;return oa(l?{"files[]":t}:t,a&&new a,this.formSerializer)}}return i||o?(n.setContentType("application/json",!1),ZN(t)):t}],transformResponse:[function(t){const n=this.transitional||If.transitional,r=n&&n.forcedJSONParsing,o=this.responseType==="json";if(t&&P.isString(t)&&(r&&!this.responseType||o)){const s=!(n&&n.silentJSONParsing)&&o;try{return JSON.parse(t)}catch(l){if(s)throw l.name==="SyntaxError"?me.from(l,me.ERR_BAD_RESPONSE,this,null,this.response):l}}return t}],timeout:0,xsrfCookieName:"XSRF-TOKEN",xsrfHeaderName:"X-XSRF-TOKEN",maxContentLength:-1,maxBodyLength:-1,env:{FormData:Zt.classes.FormData,Blob:Zt.classes.Blob},validateStatus:function(t){return t>=200&&t<300},headers:{common:{Accept:"application/json, text/plain, */*","Content-Type":void 0}}};P.forEach(["delete","get","head","post","put","patch"],e=>{If.headers[e]={}});const Vf=If,eA=P.toObjectSet(["age","authorization","content-length","content-type","etag","expires","from","host","if-modified-since","if-unmodified-since","last-modified","location","max-forwards","proxy-authorization","referer","retry-after","user-agent"]),tA=e=>{const t={};let n,r,o;return e&&e.split(`
+`).forEach(function(s){o=s.indexOf(":"),n=s.substring(0,o).trim().toLowerCase(),r=s.substring(o+1).trim(),!(!n||t[n]&&eA[n])&&(n==="set-cookie"?t[n]?t[n].push(r):t[n]=[r]:t[n]=t[n]?t[n]+", "+r:r)}),t},Jh=Symbol("internals");function Oo(e){return e&&String(e).trim().toLowerCase()}function Cs(e){return e===!1||e==null?e:P.isArray(e)?e.map(Cs):String(e)}function nA(e){const t=Object.create(null),n=/([^\s,;=]+)\s*(?:=\s*([^,;]+))?/g;let r;for(;r=n.exec(e);)t[r[1]]=r[2];return t}const rA=e=>/^[-_a-zA-Z0-9^`|~,!#$%&'*+.]+$/.test(e.trim());function Ha(e,t,n,r,o){if(P.isFunction(r))return r.call(this,t,n);if(o&&(t=n),!!P.isString(t)){if(P.isString(r))return t.indexOf(r)!==-1;if(P.isRegExp(r))return r.test(t)}}function oA(e){return e.trim().toLowerCase().replace(/([a-z\d])(\w*)/g,(t,n,r)=>n.toUpperCase()+r)}function iA(e,t){const n=P.toCamelCase(" "+t);["get","set","has"].forEach(r=>{Object.defineProperty(e,r+n,{value:function(o,i,s){return this[r].call(this,t,o,i,s)},configurable:!0})})}class ia{constructor(t){t&&this.set(t)}set(t,n,r){const o=this;function i(l,a,u){const f=Oo(a);if(!f)throw new Error("header name must be a non-empty string");const c=P.findKey(o,f);(!c||o[c]===void 0||u===!0||u===void 0&&o[c]!==!1)&&(o[c||a]=Cs(l))}const s=(l,a)=>P.forEach(l,(u,f)=>i(u,f,a));return P.isPlainObject(t)||t instanceof this.constructor?s(t,n):P.isString(t)&&(t=t.trim())&&!rA(t)?s(tA(t),n):t!=null&&i(n,t,r),this}get(t,n){if(t=Oo(t),t){const r=P.findKey(this,t);if(r){const o=this[r];if(!n)return o;if(n===!0)return nA(o);if(P.isFunction(n))return n.call(this,o,r);if(P.isRegExp(n))return n.exec(o);throw new TypeError("parser must be boolean|regexp|function")}}}has(t,n){if(t=Oo(t),t){const r=P.findKey(this,t);return!!(r&&this[r]!==void 0&&(!n||Ha(this,this[r],r,n)))}return!1}delete(t,n){const r=this;let o=!1;function i(s){if(s=Oo(s),s){const l=P.findKey(r,s);l&&(!n||Ha(r,r[l],l,n))&&(delete r[l],o=!0)}}return P.isArray(t)?t.forEach(i):i(t),o}clear(t){const n=Object.keys(this);let r=n.length,o=!1;for(;r--;){const i=n[r];(!t||Ha(this,this[i],i,t,!0))&&(delete this[i],o=!0)}return o}normalize(t){const n=this,r={};return P.forEach(this,(o,i)=>{const s=P.findKey(r,i);if(s){n[s]=Cs(o),delete n[i];return}const l=t?oA(i):String(i).trim();l!==i&&delete n[i],n[l]=Cs(o),r[l]=!0}),this}concat(...t){return this.constructor.concat(this,...t)}toJSON(t){const n=Object.create(null);return P.forEach(this,(r,o)=>{r!=null&&r!==!1&&(n[o]=t&&P.isArray(r)?r.join(", "):r)}),n}[Symbol.iterator](){return Object.entries(this.toJSON())[Symbol.iterator]()}toString(){return Object.entries(this.toJSON()).map(([t,n])=>t+": "+n).join(`
+`)}get[Symbol.toStringTag](){return"AxiosHeaders"}static from(t){return t instanceof this?t:new this(t)}static concat(t,...n){const r=new this(t);return n.forEach(o=>r.set(o)),r}static accessor(t){const r=(this[Jh]=this[Jh]={accessors:{}}).accessors,o=this.prototype;function i(s){const l=Oo(s);r[l]||(iA(o,s),r[l]=!0)}return P.isArray(t)?t.forEach(i):i(t),this}}ia.accessor(["Content-Type","Content-Length","Accept","Accept-Encoding","User-Agent","Authorization"]);P.reduceDescriptors(ia.prototype,({value:e},t)=>{let n=t[0].toUpperCase()+t.slice(1);return{get:()=>e,set(r){this[n]=r}}});P.freezeMethods(ia);const hn=ia;function Wa(e,t){const n=this||Vf,r=t||n,o=hn.from(r.headers);let i=r.data;return P.forEach(e,function(l){i=l.call(n,i,o.normalize(),t?t.status:void 0)}),o.normalize(),i}function Og(e){return!!(e&&e.__CANCEL__)}function Fi(e,t,n){me.call(this,e??"canceled",me.ERR_CANCELED,t,n),this.name="CanceledError"}P.inherits(Fi,me,{__CANCEL__:!0});function sA(e,t,n){const r=n.config.validateStatus;!n.status||!r||r(n.status)?e(n):t(new me("Request failed with status code "+n.status,[me.ERR_BAD_REQUEST,me.ERR_BAD_RESPONSE][Math.floor(n.status/100)-4],n.config,n.request,n))}const lA=Zt.hasStandardBrowserEnv?{write(e,t,n,r,o,i){const s=[e+"="+encodeURIComponent(t)];P.isNumber(n)&&s.push("expires="+new Date(n).toGMTString()),P.isString(r)&&s.push("path="+r),P.isString(o)&&s.push("domain="+o),i===!0&&s.push("secure"),document.cookie=s.join("; ")},read(e){const t=document.cookie.match(new RegExp("(^|;\\s*)("+e+")=([^;]*)"));return t?decodeURIComponent(t[3]):null},remove(e){this.write(e,"",Date.now()-864e5)}}:{write(){},read(){return null},remove(){}};function aA(e){return/^([a-z][a-z\d+\-.]*:)?\/\//i.test(e)}function uA(e,t){return t?e.replace(/\/+$/,"")+"/"+t.replace(/^\/+/,""):e}function bg(e,t){return e&&!aA(t)?uA(e,t):t}const cA=Zt.hasStandardBrowserEnv?function(){const t=/(msie|trident)/i.test(navigator.userAgent),n=document.createElement("a");let r;function o(i){let s=i;return t&&(n.setAttribute("href",s),s=n.href),n.setAttribute("href",s),{href:n.href,protocol:n.protocol?n.protocol.replace(/:$/,""):"",host:n.host,search:n.search?n.search.replace(/^\?/,""):"",hash:n.hash?n.hash.replace(/^#/,""):"",hostname:n.hostname,port:n.port,pathname:n.pathname.charAt(0)==="/"?n.pathname:"/"+n.pathname}}return r=o(window.location.href),function(s){const l=P.isString(s)?o(s):s;return l.protocol===r.protocol&&l.host===r.host}}():function(){return function(){return!0}}();function fA(e){const t=/^([-+\w]{1,25})(:?\/\/|:)/.exec(e);return t&&t[1]||""}function dA(e,t){e=e||10;const n=new Array(e),r=new Array(e);let o=0,i=0,s;return t=t!==void 0?t:1e3,function(a){const u=Date.now(),f=r[i];s||(s=u),n[o]=a,r[o]=u;let c=i,h=0;for(;c!==o;)h+=n[c++],c=c%e;if(o=(o+1)%e,o===i&&(i=(i+1)%e),u-s<t)return;const S=f&&u-f;return S?Math.round(h*1e3/S):void 0}}function Xh(e,t){let n=0;const r=dA(50,250);return o=>{const i=o.loaded,s=o.lengthComputable?o.total:void 0,l=i-n,a=r(l),u=i<=s;n=i;const f={loaded:i,total:s,progress:s?i/s:void 0,bytes:l,rate:a||void 0,estimated:a&&s&&u?(s-i)/a:void 0,event:o};f[t?"download":"upload"]=!0,e(f)}}const hA=typeof XMLHttpRequest<"u",pA=hA&&function(e){return new Promise(function(n,r){let o=e.data;const i=hn.from(e.headers).normalize();let{responseType:s,withXSRFToken:l}=e,a;function u(){e.cancelToken&&e.cancelToken.unsubscribe(a),e.signal&&e.signal.removeEventListener("abort",a)}let f;if(P.isFormData(o)){if(Zt.hasStandardBrowserEnv||Zt.hasStandardBrowserWebWorkerEnv)i.setContentType(!1);else if((f=i.getContentType())!==!1){const[v,...T]=f?f.split(";").map(m=>m.trim()).filter(Boolean):[];i.setContentType([v||"multipart/form-data",...T].join("; "))}}let c=new XMLHttpRequest;if(e.auth){const v=e.auth.username||"",T=e.auth.password?unescape(encodeURIComponent(e.auth.password)):"";i.set("Authorization","Basic "+btoa(v+":"+T))}const h=bg(e.baseURL,e.url);c.open(e.method.toUpperCase(),Ng(h,e.params,e.paramsSerializer),!0),c.timeout=e.timeout;function S(){if(!c)return;const v=hn.from("getAllResponseHeaders"in c&&c.getAllResponseHeaders()),m={data:!s||s==="text"||s==="json"?c.responseText:c.response,status:c.status,statusText:c.statusText,headers:v,config:e,request:c};sA(function(y){n(y),u()},function(y){r(y),u()},m),c=null}if("onloadend"in c?c.onloadend=S:c.onreadystatechange=function(){!c||c.readyState!==4||c.status===0&&!(c.responseURL&&c.responseURL.indexOf("file:")===0)||setTimeout(S)},c.onabort=function(){c&&(r(new me("Request aborted",me.ECONNABORTED,e,c)),c=null)},c.onerror=function(){r(new me("Network Error",me.ERR_NETWORK,e,c)),c=null},c.ontimeout=function(){let T=e.timeout?"timeout of "+e.timeout+"ms exceeded":"timeout exceeded";const m=e.transitional||Ag;e.timeoutErrorMessage&&(T=e.timeoutErrorMessage),r(new me(T,m.clarifyTimeoutError?me.ETIMEDOUT:me.ECONNABORTED,e,c)),c=null},Zt.hasStandardBrowserEnv&&(l&&P.isFunction(l)&&(l=l(e)),l||l!==!1&&cA(h))){const v=e.xsrfHeaderName&&e.xsrfCookieName&&lA.read(e.xsrfCookieName);v&&i.set(e.xsrfHeaderName,v)}o===void 0&&i.setContentType(null),"setRequestHeader"in c&&P.forEach(i.toJSON(),function(T,m){c.setRequestHeader(m,T)}),P.isUndefined(e.withCredentials)||(c.withCredentials=!!e.withCredentials),s&&s!=="json"&&(c.responseType=e.responseType),typeof e.onDownloadProgress=="function"&&c.addEventListener("progress",Xh(e.onDownloadProgress,!0)),typeof e.onUploadProgress=="function"&&c.upload&&c.upload.addEventListener("progress",Xh(e.onUploadProgress)),(e.cancelToken||e.signal)&&(a=v=>{c&&(r(!v||v.type?new Fi(null,e,c):v),c.abort(),c=null)},e.cancelToken&&e.cancelToken.subscribe(a),e.signal&&(e.signal.aborted?a():e.signal.addEventListener("abort",a)));const p=fA(h);if(p&&Zt.protocols.indexOf(p)===-1){r(new me("Unsupported protocol "+p+":",me.ERR_BAD_REQUEST,e));return}c.send(o||null)})},fc={http:IN,xhr:pA};P.forEach(fc,(e,t)=>{if(e){try{Object.defineProperty(e,"name",{value:t})}catch{}Object.defineProperty(e,"adapterName",{value:t})}});const Zh=e=>`- ${e}`,mA=e=>P.isFunction(e)||e===null||e===!1,Fg={getAdapter:e=>{e=P.isArray(e)?e:[e];const{length:t}=e;let n,r;const o={};for(let i=0;i<t;i++){n=e[i];let s;if(r=n,!mA(n)&&(r=fc[(s=String(n)).toLowerCase()],r===void 0))throw new me(`Unknown adapter '${s}'`);if(r)break;o[s||"#"+i]=r}if(!r){const i=Object.entries(o).map(([l,a])=>`adapter ${l} `+(a===!1?"is not supported by the environment":"is not available in the build"));let s=t?i.length>1?`since :
+`+i.map(Zh).join(`
+`):" "+Zh(i[0]):"as no adapter specified";throw new me("There is no suitable adapter to dispatch the request "+s,"ERR_NOT_SUPPORT")}return r},adapters:fc};function Qa(e){if(e.cancelToken&&e.cancelToken.throwIfRequested(),e.signal&&e.signal.aborted)throw new Fi(null,e)}function ep(e){return Qa(e),e.headers=hn.from(e.headers),e.data=Wa.call(e,e.transformRequest),["post","put","patch"].indexOf(e.method)!==-1&&e.headers.setContentType("application/x-www-form-urlencoded",!1),Fg.getAdapter(e.adapter||Vf.adapter)(e).then(function(r){return Qa(e),r.data=Wa.call(e,e.transformResponse,r),r.headers=hn.from(r.headers),r},function(r){return Og(r)||(Qa(e),r&&r.response&&(r.response.data=Wa.call(e,e.transformResponse,r.response),r.response.headers=hn.from(r.response.headers))),Promise.reject(r)})}const tp=e=>e instanceof hn?e.toJSON():e;function so(e,t){t=t||{};const n={};function r(u,f,c){return P.isPlainObject(u)&&P.isPlainObject(f)?P.merge.call({caseless:c},u,f):P.isPlainObject(f)?P.merge({},f):P.isArray(f)?f.slice():f}function o(u,f,c){if(P.isUndefined(f)){if(!P.isUndefined(u))return r(void 0,u,c)}else return r(u,f,c)}function i(u,f){if(!P.isUndefined(f))return r(void 0,f)}function s(u,f){if(P.isUndefined(f)){if(!P.isUndefined(u))return r(void 0,u)}else return r(void 0,f)}function l(u,f,c){if(c in t)return r(u,f);if(c in e)return r(void 0,u)}const a={url:i,method:i,data:i,baseURL:s,transformRequest:s,transformResponse:s,paramsSerializer:s,timeout:s,timeoutMessage:s,withCredentials:s,withXSRFToken:s,adapter:s,responseType:s,xsrfCookieName:s,xsrfHeaderName:s,onUploadProgress:s,onDownloadProgress:s,decompress:s,maxContentLength:s,maxBodyLength:s,beforeRedirect:s,transport:s,httpAgent:s,httpsAgent:s,cancelToken:s,socketPath:s,responseEncoding:s,validateStatus:l,headers:(u,f)=>o(tp(u),tp(f),!0)};return P.forEach(Object.keys(Object.assign({},e,t)),function(f){const c=a[f]||o,h=c(e[f],t[f],f);P.isUndefined(h)&&c!==l||(n[f]=h)}),n}const Dg="1.6.2",$f={};["object","boolean","number","function","string","symbol"].forEach((e,t)=>{$f[e]=function(r){return typeof r===e||"a"+(t<1?"n ":" ")+e}});const np={};$f.transitional=function(t,n,r){function o(i,s){return"[Axios v"+Dg+"] Transitional option '"+i+"'"+s+(r?". "+r:"")}return(i,s,l)=>{if(t===!1)throw new me(o(s," has been removed"+(n?" in "+n:"")),me.ERR_DEPRECATED);return n&&!np[s]&&(np[s]=!0,console.warn(o(s," has been deprecated since v"+n+" and will be removed in the near future"))),t?t(i,s,l):!0}};function yA(e,t,n){if(typeof e!="object")throw new me("options must be an object",me.ERR_BAD_OPTION_VALUE);const r=Object.keys(e);let o=r.length;for(;o-- >0;){const i=r[o],s=t[i];if(s){const l=e[i],a=l===void 0||s(l,i,e);if(a!==!0)throw new me("option "+i+" must be "+a,me.ERR_BAD_OPTION_VALUE);continue}if(n!==!0)throw new me("Unknown option "+i,me.ERR_BAD_OPTION)}}const dc={assertOptions:yA,validators:$f},Rn=dc.validators;class al{constructor(t){this.defaults=t,this.interceptors={request:new Yh,response:new Yh}}request(t,n){typeof t=="string"?(n=n||{},n.url=t):n=t||{},n=so(this.defaults,n);const{transitional:r,paramsSerializer:o,headers:i}=n;r!==void 0&&dc.assertOptions(r,{silentJSONParsing:Rn.transitional(Rn.boolean),forcedJSONParsing:Rn.transitional(Rn.boolean),clarifyTimeoutError:Rn.transitional(Rn.boolean)},!1),o!=null&&(P.isFunction(o)?n.paramsSerializer={serialize:o}:dc.assertOptions(o,{encode:Rn.function,serialize:Rn.function},!0)),n.method=(n.method||this.defaults.method||"get").toLowerCase();let s=i&&P.merge(i.common,i[n.method]);i&&P.forEach(["delete","get","head","post","put","patch","common"],p=>{delete i[p]}),n.headers=hn.concat(s,i);const l=[];let a=!0;this.interceptors.request.forEach(function(v){typeof v.runWhen=="function"&&v.runWhen(n)===!1||(a=a&&v.synchronous,l.unshift(v.fulfilled,v.rejected))});const u=[];this.interceptors.response.forEach(function(v){u.push(v.fulfilled,v.rejected)});let f,c=0,h;if(!a){const p=[ep.bind(this),void 0];for(p.unshift.apply(p,l),p.push.apply(p,u),h=p.length,f=Promise.resolve(n);c<h;)f=f.then(p[c++],p[c++]);return f}h=l.length;let S=n;for(c=0;c<h;){const p=l[c++],v=l[c++];try{S=p(S)}catch(T){v.call(this,T);break}}try{f=ep.call(this,S)}catch(p){return Promise.reject(p)}for(c=0,h=u.length;c<h;)f=f.then(u[c++],u[c++]);return f}getUri(t){t=so(this.defaults,t);const n=bg(t.baseURL,t.url);return Ng(n,t.params,t.paramsSerializer)}}P.forEach(["delete","get","head","options"],function(t){al.prototype[t]=function(n,r){return this.request(so(r||{},{method:t,url:n,data:(r||{}).data}))}});P.forEach(["post","put","patch"],function(t){function n(r){return function(i,s,l){return this.request(so(l||{},{method:t,headers:r?{"Content-Type":"multipart/form-data"}:{},url:i,data:s}))}}al.prototype[t]=n(),al.prototype[t+"Form"]=n(!0)});const xs=al;class jf{constructor(t){if(typeof t!="function")throw new TypeError("executor must be a function.");let n;this.promise=new Promise(function(i){n=i});const r=this;this.promise.then(o=>{if(!r._listeners)return;let i=r._listeners.length;for(;i-- >0;)r._listeners[i](o);r._listeners=null}),this.promise.then=o=>{let i;const s=new Promise(l=>{r.subscribe(l),i=l}).then(o);return s.cancel=function(){r.unsubscribe(i)},s},t(function(i,s,l){r.reason||(r.reason=new Fi(i,s,l),n(r.reason))})}throwIfRequested(){if(this.reason)throw this.reason}subscribe(t){if(this.reason){t(this.reason);return}this._listeners?this._listeners.push(t):this._listeners=[t]}unsubscribe(t){if(!this._listeners)return;const n=this._listeners.indexOf(t);n!==-1&&this._listeners.splice(n,1)}static source(){let t;return{token:new jf(function(o){t=o}),cancel:t}}}const vA=jf;function gA(e){return function(n){return e.apply(null,n)}}function SA(e){return P.isObject(e)&&e.isAxiosError===!0}const hc={Continue:100,SwitchingProtocols:101,Processing:102,EarlyHints:103,Ok:200,Created:201,Accepted:202,NonAuthoritativeInformation:203,NoContent:204,ResetContent:205,PartialContent:206,MultiStatus:207,AlreadyReported:208,ImUsed:226,MultipleChoices:300,MovedPermanently:301,Found:302,SeeOther:303,NotModified:304,UseProxy:305,Unused:306,TemporaryRedirect:307,PermanentRedirect:308,BadRequest:400,Unauthorized:401,PaymentRequired:402,Forbidden:403,NotFound:404,MethodNotAllowed:405,NotAcceptable:406,ProxyAuthenticationRequired:407,RequestTimeout:408,Conflict:409,Gone:410,LengthRequired:411,PreconditionFailed:412,PayloadTooLarge:413,UriTooLong:414,UnsupportedMediaType:415,RangeNotSatisfiable:416,ExpectationFailed:417,ImATeapot:418,MisdirectedRequest:421,UnprocessableEntity:422,Locked:423,FailedDependency:424,TooEarly:425,UpgradeRequired:426,PreconditionRequired:428,TooManyRequests:429,RequestHeaderFieldsTooLarge:431,UnavailableForLegalReasons:451,InternalServerError:500,NotImplemented:501,BadGateway:502,ServiceUnavailable:503,GatewayTimeout:504,HttpVersionNotSupported:505,VariantAlsoNegotiates:506,InsufficientStorage:507,LoopDetected:508,NotExtended:510,NetworkAuthenticationRequired:511};Object.entries(hc).forEach(([e,t])=>{hc[t]=e});const wA=hc;function Mg(e){const t=new xs(e),n=yg(xs.prototype.request,t);return P.extend(n,xs.prototype,t,{allOwnKeys:!0}),P.extend(n,t,null,{allOwnKeys:!0}),n.create=function(o){return Mg(so(e,o))},n}const We=Mg(Vf);We.Axios=xs;We.CanceledError=Fi;We.CancelToken=vA;We.isCancel=Og;We.VERSION=Dg;We.toFormData=oa;We.AxiosError=me;We.Cancel=We.CanceledError;We.all=function(t){return Promise.all(t)};We.spread=gA;We.isAxiosError=SA;We.mergeConfig=so;We.AxiosHeaders=hn;We.formToJSON=e=>Pg(P.isHTMLForm(e)?new FormData(e):e);We.getAdapter=Fg.getAdapter;We.HttpStatusCode=wA;We.default=We;const Ug=We,Zo=Ug.create({baseURL:"/services/japps",headers:{"Content-Type":"application/json"},params:{token:"13748105b2e64b9082e10f802b04f523"}}),_A="This field is required.",rp={required:_A},Bf=()=>window.jhdata,RA=(e,t)=>{var r;const n=[];for(const o in e)if(Object.hasOwnProperty.call(e,o)){const i=e[o];i.display===!0&&n.push({name:i.info.name,url:(r=i.info.url)==null?void 0:r.replace("[USER]",t),external:i.info.external})}return n},EA=(e,t)=>{var r;const n=[];for(const o in e)if(Object.hasOwnProperty.call(e,o)){const i=e[o];if((r=i.user_options)!=null&&r.jhub_app){const s=i.user_options;n.push({id:s.name,name:s.display_name,description:s.description,framework:CA(s.framework),url:i.url,thumbnail:s.thumbnail,shared:!1,ready:s.ready})}}return t.toLowerCase()==="shared"?n.filter(o=>o.shared===!0):n.filter(o=>o.shared===!1)},CA=e=>e.charAt(0).toUpperCase()+e.slice(1);var Di=e=>e.type==="checkbox",Ur=e=>e instanceof Date,at=e=>e==null;const Ig=e=>typeof e=="object";var Ke=e=>!at(e)&&!Array.isArray(e)&&Ig(e)&&!Ur(e),Vg=e=>Ke(e)&&e.target?Di(e.target)?e.target.checked:e.target.value:e,xA=e=>e.substring(0,e.search(/\.\d+(\.|$)/))||e,$g=(e,t)=>e.has(xA(t)),TA=e=>{const t=e.constructor&&e.constructor.prototype;return Ke(t)&&t.hasOwnProperty("isPrototypeOf")},zf=typeof window<"u"&&typeof window.HTMLElement<"u"&&typeof document<"u";function Bt(e){let t;const n=Array.isArray(e);if(e instanceof Date)t=new Date(e);else if(e instanceof Set)t=new Set(e);else if(!(zf&&(e instanceof Blob||e instanceof FileList))&&(n||Ke(e)))if(t=n?[]:{},!n&&!TA(e))t=e;else for(const r in e)e.hasOwnProperty(r)&&(t[r]=Bt(e[r]));else return e;return t}var Mi=e=>Array.isArray(e)?e.filter(Boolean):[],je=e=>e===void 0,W=(e,t,n)=>{if(!t||!Ke(e))return n;const r=Mi(t.split(/[,[\].]+?/)).reduce((o,i)=>at(o)?o:o[i],e);return je(r)||r===e?je(e[t])?n:e[t]:r},On=e=>typeof e=="boolean";const ul={BLUR:"blur",FOCUS_OUT:"focusout",CHANGE:"change"},Wt={onBlur:"onBlur",onChange:"onChange",onSubmit:"onSubmit",onTouched:"onTouched",all:"all"},ln={max:"max",min:"min",maxLength:"maxLength",minLength:"minLength",pattern:"pattern",required:"required",validate:"validate"},kA=ne.createContext(null),Hf=()=>ne.useContext(kA);var jg=(e,t,n,r=!0)=>{const o={defaultValues:t._defaultValues};for(const i in e)Object.defineProperty(o,i,{get:()=>{const s=i;return t._proxyFormState[s]!==Wt.all&&(t._proxyFormState[s]=!r||Wt.all),n&&(n[s]=!0),e[s]}});return o},Lt=e=>Ke(e)&&!Object.keys(e).length,Bg=(e,t,n,r)=>{n(e);const{name:o,...i}=e;return Lt(i)||Object.keys(i).length>=Object.keys(t).length||Object.keys(i).find(s=>t[s]===(!r||Wt.all))},Ts=e=>Array.isArray(e)?e:[e],zg=(e,t,n)=>!e||!t||e===t||Ts(e).some(r=>r&&(n?r===t:r.startsWith(t)||t.startsWith(r)));function Wf(e){const t=ne.useRef(e);t.current=e,ne.useEffect(()=>{const n=!e.disabled&&t.current.subject&&t.current.subject.subscribe({next:t.current.next});return()=>{n&&n.unsubscribe()}},[e.disabled])}function NA(e){const t=Hf(),{control:n=t.control,disabled:r,name:o,exact:i}=e||{},[s,l]=ne.useState(n._formState),a=ne.useRef(!0),u=ne.useRef({isDirty:!1,isLoading:!1,dirtyFields:!1,touchedFields:!1,isValidating:!1,isValid:!1,errors:!1}),f=ne.useRef(o);return f.current=o,Wf({disabled:r,next:c=>a.current&&zg(f.current,c.name,i)&&Bg(c,u.current,n._updateFormState)&&l({...n._formState,...c}),subject:n._subjects.state}),ne.useEffect(()=>(a.current=!0,u.current.isValid&&n._updateValid(!0),()=>{a.current=!1}),[n]),jg(s,n,u.current,!1)}var en=e=>typeof e=="string",Hg=(e,t,n,r,o)=>en(e)?(r&&t.watch.add(e),W(n,e,o)):Array.isArray(e)?e.map(i=>(r&&t.watch.add(i),W(n,i))):(r&&(t.watchAll=!0),n);function AA(e){const t=Hf(),{control:n=t.control,name:r,defaultValue:o,disabled:i,exact:s}=e||{},l=ne.useRef(r);l.current=r,Wf({disabled:i,subject:n._subjects.values,next:f=>{zg(l.current,f.name,s)&&u(Bt(Hg(l.current,n._names,f.values||n._formValues,!1,o)))}});const[a,u]=ne.useState(n._getWatch(r,o));return ne.useEffect(()=>n._removeUnmounted()),a}var Qf=e=>/^\w*$/.test(e),Wg=e=>Mi(e.replace(/["|']|\]/g,"").split(/\.|\[/));function ke(e,t,n){let r=-1;const o=Qf(t)?[t]:Wg(t),i=o.length,s=i-1;for(;++r<i;){const l=o[r];let a=n;if(r!==s){const u=e[l];a=Ke(u)||Array.isArray(u)?u:isNaN(+o[r+1])?{}:[]}e[l]=a,e=e[l]}return e}function LA(e){const t=Hf(),{name:n,disabled:r,control:o=t.control,shouldUnregister:i}=e,s=$g(o._names.array,n),l=AA({control:o,name:n,defaultValue:W(o._formValues,n,W(o._defaultValues,n,e.defaultValue)),exact:!0}),a=NA({control:o,name:n}),u=ne.useRef(o.register(n,{...e.rules,value:l}));return u.current=o.register(n,e.rules),ne.useEffect(()=>{const f=o._options.shouldUnregister||i,c=(h,S)=>{const p=W(o._fields,h);p&&(p._f.mount=S)};if(c(n,!0),f){const h=Bt(W(o._options.defaultValues,n));ke(o._defaultValues,n,h),je(W(o._formValues,n))&&ke(o._formValues,n,h)}return()=>{(s?f&&!o._state.action:f)?o.unregister(n):c(n,!1)}},[n,o,s,i]),ne.useEffect(()=>{W(o._fields,n)&&o._updateDisabledField({disabled:r,fields:o._fields,name:n})},[r,n,o]),{field:{name:n,value:l,...On(r)?{disabled:r}:{},onChange:ne.useCallback(f=>u.current.onChange({target:{value:Vg(f),name:n},type:ul.CHANGE}),[n]),onBlur:ne.useCallback(()=>u.current.onBlur({target:{value:W(o._formValues,n),name:n},type:ul.BLUR}),[n,o]),ref:f=>{const c=W(o._fields,n);c&&f&&(c._f.ref={focus:()=>f.focus(),select:()=>f.select(),setCustomValidity:h=>f.setCustomValidity(h),reportValidity:()=>f.reportValidity()})}},formState:a,fieldState:Object.defineProperties({},{invalid:{enumerable:!0,get:()=>!!W(a.errors,n)},isDirty:{enumerable:!0,get:()=>!!W(a.dirtyFields,n)},isTouched:{enumerable:!0,get:()=>!!W(a.touchedFields,n)},error:{enumerable:!0,get:()=>W(a.errors,n)}})}}const Er=e=>e.render(LA(e));var PA=(e,t,n,r,o)=>t?{...n[e],types:{...n[e]&&n[e].types?n[e].types:{},[r]:o||!0}}:{};const pc=(e,t,n)=>{for(const r of n||Object.keys(e)){const o=W(e,r);if(o){const{_f:i,...s}=o;if(i&&t(i.name)){if(i.ref.focus){i.ref.focus();break}else if(i.refs&&i.refs[0].focus){i.refs[0].focus();break}}else Ke(s)&&pc(s,t)}}};var op=e=>({isOnSubmit:!e||e===Wt.onSubmit,isOnBlur:e===Wt.onBlur,isOnChange:e===Wt.onChange,isOnAll:e===Wt.all,isOnTouch:e===Wt.onTouched}),ip=(e,t,n)=>!n&&(t.watchAll||t.watch.has(e)||[...t.watch].some(r=>e.startsWith(r)&&/^\.\w+/.test(e.slice(r.length)))),OA=(e,t,n)=>{const r=Mi(W(e,n));return ke(r,"root",t[n]),ke(e,n,r),e},qf=e=>e.type==="file",bn=e=>typeof e=="function",cl=e=>{if(!zf)return!1;const t=e?e.ownerDocument:0;return e instanceof(t&&t.defaultView?t.defaultView.HTMLElement:HTMLElement)},ks=e=>en(e),Kf=e=>e.type==="radio",fl=e=>e instanceof RegExp;const sp={value:!1,isValid:!1},lp={value:!0,isValid:!0};var Qg=e=>{if(Array.isArray(e)){if(e.length>1){const t=e.filter(n=>n&&n.checked&&!n.disabled).map(n=>n.value);return{value:t,isValid:!!t.length}}return e[0].checked&&!e[0].disabled?e[0].attributes&&!je(e[0].attributes.value)?je(e[0].value)||e[0].value===""?lp:{value:e[0].value,isValid:!0}:lp:sp}return sp};const ap={isValid:!1,value:null};var qg=e=>Array.isArray(e)?e.reduce((t,n)=>n&&n.checked&&!n.disabled?{isValid:!0,value:n.value}:t,ap):ap;function up(e,t,n="validate"){if(ks(e)||Array.isArray(e)&&e.every(ks)||On(e)&&!e)return{type:n,message:ks(e)?e:"",ref:t}}var Cr=e=>Ke(e)&&!fl(e)?e:{value:e,message:""},cp=async(e,t,n,r,o)=>{const{ref:i,refs:s,required:l,maxLength:a,minLength:u,min:f,max:c,pattern:h,validate:S,name:p,valueAsNumber:v,mount:T,disabled:m}=e._f,d=W(t,p);if(!T||m)return{};const y=s?s[0]:i,_=ae=>{r&&y.reportValidity&&(y.setCustomValidity(On(ae)?"":ae||""),y.reportValidity())},A={},k=Kf(i),x=Di(i),O=k||x,J=(v||qf(i))&&je(i.value)&&je(d)||cl(i)&&i.value===""||d===""||Array.isArray(d)&&!d.length,G=PA.bind(null,p,n,A),ye=(ae,B,ie,pe=ln.maxLength,ge=ln.minLength)=>{const _e=ae?B:ie;A[p]={type:ae?pe:ge,message:_e,ref:i,...G(ae?pe:ge,_e)}};if(o?!Array.isArray(d)||!d.length:l&&(!O&&(J||at(d))||On(d)&&!d||x&&!Qg(s).isValid||k&&!qg(s).isValid)){const{value:ae,message:B}=ks(l)?{value:!!l,message:l}:Cr(l);if(ae&&(A[p]={type:ln.required,message:B,ref:y,...G(ln.required,B)},!n))return _(B),A}if(!J&&(!at(f)||!at(c))){let ae,B;const ie=Cr(c),pe=Cr(f);if(!at(d)&&!isNaN(d)){const ge=i.valueAsNumber||d&&+d;at(ie.value)||(ae=ge>ie.value),at(pe.value)||(B=ge<pe.value)}else{const ge=i.valueAsDate||new Date(d),_e=te=>new Date(new Date().toDateString()+" "+te),V=i.type=="time",X=i.type=="week";en(ie.value)&&d&&(ae=V?_e(d)>_e(ie.value):X?d>ie.value:ge>new Date(ie.value)),en(pe.value)&&d&&(B=V?_e(d)<_e(pe.value):X?d<pe.value:ge<new Date(pe.value))}if((ae||B)&&(ye(!!ae,ie.message,pe.message,ln.max,ln.min),!n))return _(A[p].message),A}if((a||u)&&!J&&(en(d)||o&&Array.isArray(d))){const ae=Cr(a),B=Cr(u),ie=!at(ae.value)&&d.length>+ae.value,pe=!at(B.value)&&d.length<+B.value;if((ie||pe)&&(ye(ie,ae.message,B.message),!n))return _(A[p].message),A}if(h&&!J&&en(d)){const{value:ae,message:B}=Cr(h);if(fl(ae)&&!d.match(ae)&&(A[p]={type:ln.pattern,message:B,ref:i,...G(ln.pattern,B)},!n))return _(B),A}if(S){if(bn(S)){const ae=await S(d,t),B=up(ae,y);if(B&&(A[p]={...B,...G(ln.validate,B.message)},!n))return _(B.message),A}else if(Ke(S)){let ae={};for(const B in S){if(!Lt(ae)&&!n)break;const ie=up(await S[B](d,t),y,B);ie&&(ae={...ie,...G(B,ie.message)},_(ie.message),n&&(A[p]=ae))}if(!Lt(ae)&&(A[p]={ref:y,...ae},!n))return A}}return _(!0),A};function bA(e,t){const n=t.slice(0,-1).length;let r=0;for(;r<n;)e=je(e)?r++:e[t[r++]];return e}function FA(e){for(const t in e)if(e.hasOwnProperty(t)&&!je(e[t]))return!1;return!0}function Xe(e,t){const n=Array.isArray(t)?t:Qf(t)?[t]:Wg(t),r=n.length===1?e:bA(e,n),o=n.length-1,i=n[o];return r&&delete r[i],o!==0&&(Ke(r)&&Lt(r)||Array.isArray(r)&&FA(r))&&Xe(e,n.slice(0,-1)),e}function qa(){let e=[];return{get observers(){return e},next:o=>{for(const i of e)i.next&&i.next(o)},subscribe:o=>(e.push(o),{unsubscribe:()=>{e=e.filter(i=>i!==o)}}),unsubscribe:()=>{e=[]}}}var dl=e=>at(e)||!Ig(e);function or(e,t){if(dl(e)||dl(t))return e===t;if(Ur(e)&&Ur(t))return e.getTime()===t.getTime();const n=Object.keys(e),r=Object.keys(t);if(n.length!==r.length)return!1;for(const o of n){const i=e[o];if(!r.includes(o))return!1;if(o!=="ref"){const s=t[o];if(Ur(i)&&Ur(s)||Ke(i)&&Ke(s)||Array.isArray(i)&&Array.isArray(s)?!or(i,s):i!==s)return!1}}return!0}var Kg=e=>e.type==="select-multiple",DA=e=>Kf(e)||Di(e),Ka=e=>cl(e)&&e.isConnected,Gg=e=>{for(const t in e)if(bn(e[t]))return!0;return!1};function hl(e,t={}){const n=Array.isArray(e);if(Ke(e)||n)for(const r in e)Array.isArray(e[r])||Ke(e[r])&&!Gg(e[r])?(t[r]=Array.isArray(e[r])?[]:{},hl(e[r],t[r])):at(e[r])||(t[r]=!0);return t}function Yg(e,t,n){const r=Array.isArray(e);if(Ke(e)||r)for(const o in e)Array.isArray(e[o])||Ke(e[o])&&!Gg(e[o])?je(t)||dl(n[o])?n[o]=Array.isArray(e[o])?hl(e[o],[]):{...hl(e[o])}:Yg(e[o],at(t)?{}:t[o],n[o]):n[o]=!or(e[o],t[o]);return n}var Ga=(e,t)=>Yg(e,t,hl(t)),Jg=(e,{valueAsNumber:t,valueAsDate:n,setValueAs:r})=>je(e)?e:t?e===""?NaN:e&&+e:n&&en(e)?new Date(e):r?r(e):e;function Ya(e){const t=e.ref;if(!(e.refs?e.refs.every(n=>n.disabled):t.disabled))return qf(t)?t.files:Kf(t)?qg(e.refs).value:Kg(t)?[...t.selectedOptions].map(({value:n})=>n):Di(t)?Qg(e.refs).value:Jg(je(t.value)?e.ref.value:t.value,e)}var MA=(e,t,n,r)=>{const o={};for(const i of e){const s=W(t,i);s&&ke(o,i,s._f)}return{criteriaMode:n,names:[...e],fields:o,shouldUseNativeValidation:r}},bo=e=>je(e)?e:fl(e)?e.source:Ke(e)?fl(e.value)?e.value.source:e.value:e,UA=e=>e.mount&&(e.required||e.min||e.max||e.maxLength||e.minLength||e.pattern||e.validate);function fp(e,t,n){const r=W(e,n);if(r||Qf(n))return{error:r,name:n};const o=n.split(".");for(;o.length;){const i=o.join("."),s=W(t,i),l=W(e,i);if(s&&!Array.isArray(s)&&n!==i)return{name:n};if(l&&l.type)return{name:i,error:l};o.pop()}return{name:n}}var IA=(e,t,n,r,o)=>o.isOnAll?!1:!n&&o.isOnTouch?!(t||e):(n?r.isOnBlur:o.isOnBlur)?!e:(n?r.isOnChange:o.isOnChange)?e:!0,VA=(e,t)=>!Mi(W(e,t)).length&&Xe(e,t);const $A={mode:Wt.onSubmit,reValidateMode:Wt.onChange,shouldFocusError:!0};function jA(e={},t){let n={...$A,...e},r={submitCount:0,isDirty:!1,isLoading:bn(n.defaultValues),isValidating:!1,isSubmitted:!1,isSubmitting:!1,isSubmitSuccessful:!1,isValid:!1,touchedFields:{},dirtyFields:{},errors:{}},o={},i=Ke(n.defaultValues)||Ke(n.values)?Bt(n.defaultValues||n.values)||{}:{},s=n.shouldUnregister?{}:Bt(i),l={action:!1,mount:!1,watch:!1},a={mount:new Set,unMount:new Set,array:new Set,watch:new Set},u,f=0;const c={isDirty:!1,dirtyFields:!1,touchedFields:!1,isValidating:!1,isValid:!1,errors:!1},h={values:qa(),array:qa(),state:qa()},S=e.resetOptions&&e.resetOptions.keepDirtyValues,p=op(n.mode),v=op(n.reValidateMode),T=n.criteriaMode===Wt.all,m=g=>N=>{clearTimeout(f),f=setTimeout(g,N)},d=async g=>{if(c.isValid||g){const N=n.resolver?Lt((await J()).errors):await ye(o,!0);N!==r.isValid&&h.state.next({isValid:N})}},y=g=>c.isValidating&&h.state.next({isValidating:g}),_=(g,N=[],L,K,$=!0,I=!0)=>{if(K&&L){if(l.action=!0,I&&Array.isArray(W(o,g))){const ee=L(W(o,g),K.argA,K.argB);$&&ke(o,g,ee)}if(I&&Array.isArray(W(r.errors,g))){const ee=L(W(r.errors,g),K.argA,K.argB);$&&ke(r.errors,g,ee),VA(r.errors,g)}if(c.touchedFields&&I&&Array.isArray(W(r.touchedFields,g))){const ee=L(W(r.touchedFields,g),K.argA,K.argB);$&&ke(r.touchedFields,g,ee)}c.dirtyFields&&(r.dirtyFields=Ga(i,s)),h.state.next({name:g,isDirty:B(g,N),dirtyFields:r.dirtyFields,errors:r.errors,isValid:r.isValid})}else ke(s,g,N)},A=(g,N)=>{ke(r.errors,g,N),h.state.next({errors:r.errors})},k=(g,N,L,K)=>{const $=W(o,g);if($){const I=W(s,g,je(L)?W(i,g):L);je(I)||K&&K.defaultChecked||N?ke(s,g,N?I:Ya($._f)):ge(g,I),l.mount&&d()}},x=(g,N,L,K,$)=>{let I=!1,ee=!1;const Te={name:g};if(!L||K){c.isDirty&&(ee=r.isDirty,r.isDirty=Te.isDirty=B(),I=ee!==Te.isDirty);const Ae=or(W(i,g),N);ee=W(r.dirtyFields,g),Ae?Xe(r.dirtyFields,g):ke(r.dirtyFields,g,!0),Te.dirtyFields=r.dirtyFields,I=I||c.dirtyFields&&ee!==!Ae}if(L){const Ae=W(r.touchedFields,g);Ae||(ke(r.touchedFields,g,L),Te.touchedFields=r.touchedFields,I=I||c.touchedFields&&Ae!==L)}return I&&$&&h.state.next(Te),I?Te:{}},O=(g,N,L,K)=>{const $=W(r.errors,g),I=c.isValid&&On(N)&&r.isValid!==N;if(e.delayError&&L?(u=m(()=>A(g,L)),u(e.delayError)):(clearTimeout(f),u=null,L?ke(r.errors,g,L):Xe(r.errors,g)),(L?!or($,L):$)||!Lt(K)||I){const ee={...K,...I&&On(N)?{isValid:N}:{},errors:r.errors,name:g};r={...r,...ee},h.state.next(ee)}y(!1)},J=async g=>n.resolver(s,n.context,MA(g||a.mount,o,n.criteriaMode,n.shouldUseNativeValidation)),G=async g=>{const{errors:N}=await J(g);if(g)for(const L of g){const K=W(N,L);K?ke(r.errors,L,K):Xe(r.errors,L)}else r.errors=N;return N},ye=async(g,N,L={valid:!0})=>{for(const K in g){const $=g[K];if($){const{_f:I,...ee}=$;if(I){const Te=a.array.has(I.name),Ae=await cp($,s,T,n.shouldUseNativeValidation&&!N,Te);if(Ae[I.name]&&(L.valid=!1,N))break;!N&&(W(Ae,I.name)?Te?OA(r.errors,Ae,I.name):ke(r.errors,I.name,Ae[I.name]):Xe(r.errors,I.name))}ee&&await ye(ee,N,L)}}return L.valid},ae=()=>{for(const g of a.unMount){const N=W(o,g);N&&(N._f.refs?N._f.refs.every(L=>!Ka(L)):!Ka(N._f.ref))&&j(g)}a.unMount=new Set},B=(g,N)=>(g&&N&&ke(s,g,N),!or(Se(),i)),ie=(g,N,L)=>Hg(g,a,{...l.mount?s:je(N)?i:en(g)?{[g]:N}:N},L,N),pe=g=>Mi(W(l.mount?s:i,g,e.shouldUnregister?W(i,g,[]):[])),ge=(g,N,L={})=>{const K=W(o,g);let $=N;if(K){const I=K._f;I&&(!I.disabled&&ke(s,g,Jg(N,I)),$=cl(I.ref)&&at(N)?"":N,Kg(I.ref)?[...I.ref.options].forEach(ee=>ee.selected=$.includes(ee.value)):I.refs?Di(I.ref)?I.refs.length>1?I.refs.forEach(ee=>(!ee.defaultChecked||!ee.disabled)&&(ee.checked=Array.isArray($)?!!$.find(Te=>Te===ee.value):$===ee.value)):I.refs[0]&&(I.refs[0].checked=!!$):I.refs.forEach(ee=>ee.checked=ee.value===$):qf(I.ref)?I.ref.value="":(I.ref.value=$,I.ref.type||h.values.next({name:g,values:{...s}})))}(L.shouldDirty||L.shouldTouch)&&x(g,$,L.shouldTouch,L.shouldDirty,!0),L.shouldValidate&&te(g)},_e=(g,N,L)=>{for(const K in N){const $=N[K],I=`${g}.${K}`,ee=W(o,I);(a.array.has(g)||!dl($)||ee&&!ee._f)&&!Ur($)?_e(I,$,L):ge(I,$,L)}},V=(g,N,L={})=>{const K=W(o,g),$=a.array.has(g),I=Bt(N);ke(s,g,I),$?(h.array.next({name:g,values:{...s}}),(c.isDirty||c.dirtyFields)&&L.shouldDirty&&h.state.next({name:g,dirtyFields:Ga(i,s),isDirty:B(g,I)})):K&&!K._f&&!at(I)?_e(g,I,L):ge(g,I,L),ip(g,a)&&h.state.next({...r}),h.values.next({name:g,values:{...s}}),!l.mount&&t()},X=async g=>{const N=g.target;let L=N.name,K=!0;const $=W(o,L),I=()=>N.type?Ya($._f):Vg(g);if($){let ee,Te;const Ae=I(),wn=g.type===ul.BLUR||g.type===ul.FOCUS_OUT,sa=!UA($._f)&&!n.resolver&&!W(r.errors,L)&&!$._f.deps||IA(wn,W(r.touchedFields,L),r.isSubmitted,v,p),mo=ip(L,a,wn);ke(s,L,Ae),wn?($._f.onBlur&&$._f.onBlur(g),u&&u(0)):$._f.onChange&&$._f.onChange(g);const yo=x(L,Ae,wn,!1),la=!Lt(yo)||mo;if(!wn&&h.values.next({name:L,type:g.type,values:{...s}}),sa)return c.isValid&&d(),la&&h.state.next({name:L,...mo?{}:yo});if(!wn&&mo&&h.state.next({...r}),y(!0),n.resolver){const{errors:Ui}=await J([L]),Ii=fp(r.errors,o,L),Vi=fp(Ui,o,Ii.name||L);ee=Vi.error,L=Vi.name,Te=Lt(Ui)}else ee=(await cp($,s,T,n.shouldUseNativeValidation))[L],K=Number.isNaN(Ae)||Ae===W(s,L,Ae),K&&(ee?Te=!1:c.isValid&&(Te=await ye(o,!0)));K&&($._f.deps&&te($._f.deps),O(L,Te,ee,yo))}},te=async(g,N={})=>{let L,K;const $=Ts(g);if(y(!0),n.resolver){const I=await G(je(g)?g:$);L=Lt(I),K=g?!$.some(ee=>W(I,ee)):L}else g?(K=(await Promise.all($.map(async I=>{const ee=W(o,I);return await ye(ee&&ee._f?{[I]:ee}:ee)}))).every(Boolean),!(!K&&!r.isValid)&&d()):K=L=await ye(o);return h.state.next({...!en(g)||c.isValid&&L!==r.isValid?{}:{name:g},...n.resolver||!g?{isValid:L}:{},errors:r.errors,isValidating:!1}),N.shouldFocus&&!K&&pc(o,I=>I&&W(r.errors,I),g?$:a.mount),K},Se=g=>{const N={...i,...l.mount?s:{}};return je(g)?N:en(g)?W(N,g):g.map(L=>W(N,L))},E=(g,N)=>({invalid:!!W((N||r).errors,g),isDirty:!!W((N||r).dirtyFields,g),isTouched:!!W((N||r).touchedFields,g),error:W((N||r).errors,g)}),b=g=>{g&&Ts(g).forEach(N=>Xe(r.errors,N)),h.state.next({errors:g?r.errors:{}})},F=(g,N,L)=>{const K=(W(o,g,{_f:{}})._f||{}).ref;ke(r.errors,g,{...N,ref:K}),h.state.next({name:g,errors:r.errors,isValid:!1}),L&&L.shouldFocus&&K&&K.focus&&K.focus()},Y=(g,N)=>bn(g)?h.values.subscribe({next:L=>g(ie(void 0,N),L)}):ie(g,N,!0),j=(g,N={})=>{for(const L of g?Ts(g):a.mount)a.mount.delete(L),a.array.delete(L),N.keepValue||(Xe(o,L),Xe(s,L)),!N.keepError&&Xe(r.errors,L),!N.keepDirty&&Xe(r.dirtyFields,L),!N.keepTouched&&Xe(r.touchedFields,L),!n.shouldUnregister&&!N.keepDefaultValue&&Xe(i,L);h.values.next({values:{...s}}),h.state.next({...r,...N.keepDirty?{isDirty:B()}:{}}),!N.keepIsValid&&d()},le=({disabled:g,name:N,field:L,fields:K})=>{if(On(g)){const $=g?void 0:W(s,N,Ya(L?L._f:W(K,N)._f));ke(s,N,$),x(N,$,!1,!1,!0)}},ue=(g,N={})=>{let L=W(o,g);const K=On(N.disabled);return ke(o,g,{...L||{},_f:{...L&&L._f?L._f:{ref:{name:g}},name:g,mount:!0,...N}}),a.mount.add(g),L?le({field:L,disabled:N.disabled,name:g}):k(g,!0,N.value),{...K?{disabled:N.disabled}:{},...n.progressive?{required:!!N.required,min:bo(N.min),max:bo(N.max),minLength:bo(N.minLength),maxLength:bo(N.maxLength),pattern:bo(N.pattern)}:{},name:g,onChange:X,onBlur:X,ref:$=>{if($){ue(g,N),L=W(o,g);const I=je($.value)&&$.querySelectorAll&&$.querySelectorAll("input,select,textarea")[0]||$,ee=DA(I),Te=L._f.refs||[];if(ee?Te.find(Ae=>Ae===I):I===L._f.ref)return;ke(o,g,{_f:{...L._f,...ee?{refs:[...Te.filter(Ka),I,...Array.isArray(W(i,g))?[{}]:[]],ref:{type:I.type,name:g}}:{ref:I}}}),k(g,!1,void 0,I)}else L=W(o,g,{}),L._f&&(L._f.mount=!1),(n.shouldUnregister||N.shouldUnregister)&&!($g(a.array,g)&&l.action)&&a.unMount.add(g)}}},Z=()=>n.shouldFocusError&&pc(o,g=>g&&W(r.errors,g),a.mount),Ee=(g,N)=>async L=>{L&&(L.preventDefault&&L.preventDefault(),L.persist&&L.persist());let K=Bt(s);if(h.state.next({isSubmitting:!0}),n.resolver){const{errors:$,values:I}=await J();r.errors=$,K=I}else await ye(o);Xe(r.errors,"root"),Lt(r.errors)?(h.state.next({errors:{}}),await g(K,L)):(N&&await N({...r.errors},L),Z(),setTimeout(Z)),h.state.next({isSubmitted:!0,isSubmitting:!1,isSubmitSuccessful:Lt(r.errors),submitCount:r.submitCount+1,errors:r.errors})},ze=(g,N={})=>{W(o,g)&&(je(N.defaultValue)?V(g,W(i,g)):(V(g,N.defaultValue),ke(i,g,N.defaultValue)),N.keepTouched||Xe(r.touchedFields,g),N.keepDirty||(Xe(r.dirtyFields,g),r.isDirty=N.defaultValue?B(g,W(i,g)):B()),N.keepError||(Xe(r.errors,g),c.isValid&&d()),h.state.next({...r}))},we=(g,N={})=>{const L=g?Bt(g):i,K=Bt(L),$=g&&!Lt(g)?K:i;if(N.keepDefaultValues||(i=L),!N.keepValues){if(N.keepDirtyValues||S)for(const I of a.mount)W(r.dirtyFields,I)?ke($,I,W(s,I)):V(I,W($,I));else{if(zf&&je(g))for(const I of a.mount){const ee=W(o,I);if(ee&&ee._f){const Te=Array.isArray(ee._f.refs)?ee._f.refs[0]:ee._f.ref;if(cl(Te)){const Ae=Te.closest("form");if(Ae){Ae.reset();break}}}}o={}}s=e.shouldUnregister?N.keepDefaultValues?Bt(i):{}:Bt($),h.array.next({values:{...$}}),h.values.next({values:{...$}})}a={mount:new Set,unMount:new Set,array:new Set,watch:new Set,watchAll:!1,focus:""},!l.mount&&t(),l.mount=!c.isValid||!!N.keepIsValid,l.watch=!!e.shouldUnregister,h.state.next({submitCount:N.keepSubmitCount?r.submitCount:0,isDirty:N.keepDirty?r.isDirty:!!(N.keepDefaultValues&&!or(g,i)),isSubmitted:N.keepIsSubmitted?r.isSubmitted:!1,dirtyFields:N.keepDirtyValues?r.dirtyFields:N.keepDefaultValues&&g?Ga(i,g):{},touchedFields:N.keepTouched?r.touchedFields:{},errors:N.keepErrors?r.errors:{},isSubmitSuccessful:N.keepIsSubmitSuccessful?r.isSubmitSuccessful:!1,isSubmitting:!1})},Ce=(g,N)=>we(bn(g)?g(s):g,N);return{control:{register:ue,unregister:j,getFieldState:E,handleSubmit:Ee,setError:F,_executeSchema:J,_getWatch:ie,_getDirty:B,_updateValid:d,_removeUnmounted:ae,_updateFieldArray:_,_updateDisabledField:le,_getFieldArray:pe,_reset:we,_resetDefaultValues:()=>bn(n.defaultValues)&&n.defaultValues().then(g=>{Ce(g,n.resetOptions),h.state.next({isLoading:!1})}),_updateFormState:g=>{r={...r,...g}},_subjects:h,_proxyFormState:c,get _fields(){return o},get _formValues(){return s},get _state(){return l},set _state(g){l=g},get _defaultValues(){return i},get _names(){return a},set _names(g){a=g},get _formState(){return r},set _formState(g){r=g},get _options(){return n},set _options(g){n={...n,...g}}},trigger:te,register:ue,handleSubmit:Ee,watch:Y,setValue:V,getValues:Se,reset:Ce,resetField:ze,clearErrors:b,unregister:j,setError:F,setFocus:(g,N={})=>{const L=W(o,g),K=L&&L._f;if(K){const $=K.refs?K.refs[0]:K.ref;$.focus&&($.focus(),N.shouldSelect&&$.select())}},getFieldState:E}}function BA(e={}){const t=ne.useRef(),n=ne.useRef(),[r,o]=ne.useState({isDirty:!1,isValidating:!1,isLoading:bn(e.defaultValues),isSubmitted:!1,isSubmitting:!1,isSubmitSuccessful:!1,isValid:!1,submitCount:0,dirtyFields:{},touchedFields:{},errors:{},defaultValues:bn(e.defaultValues)?void 0:e.defaultValues});t.current||(t.current={...jA(e,()=>o(s=>({...s}))),formState:r});const i=t.current.control;return i._options=e,Wf({subject:i._subjects.state,next:s=>{Bg(s,i._proxyFormState,i._updateFormState,!0)&&o({...i._formState})}}),ne.useEffect(()=>{e.values&&!or(e.values,n.current)?(i._reset(e.values,i._options.resetOptions),n.current=e.values):i._resetDefaultValues()},[e.values,i]),ne.useEffect(()=>{i._state.mount||(i._updateValid(),i._state.mount=!0),i._state.watch&&(i._state.watch=!1,i._subjects.state.next({...i._formState})),i._removeUnmounted()}),t.current.formState=jg(r,i),t.current}const Xg=({id:e,onCancel:t,onSubmit:n})=>{var _,A;const r=ea(),[o,i]=U.useState(!1),[,s]=Hn(Oi),[l,a]=U.useState(""),{data:u,error:f}=Ff({queryKey:["app-form",e],queryFn:()=>Zo.get(`/server/${e}`).then(k=>k.data),enabled:!!e}),{control:c,handleSubmit:h,reset:S,formState:{errors:p}}=BA({defaultValues:{display_name:"",description:"",framework:"",thumbnail:"",filepath:"",conda_env:"",custom_command:"",profile:""}}),v=({display_name:k,description:x,framework:O,thumbnail:J,filepath:G,conda_env:ye,custom_command:ae,profile:B})=>{const ie={servername:l||k,user_options:{jhub_app:!0,name:l||k,display_name:k,description:x||"",framework:O,thumbnail:J||"",filepath:G||"",conda_env:ye||"",custom_command:ae||"",profile:B||""}};i(!0),e?y(ie,{onSuccess:async()=>{i(!1),r.invalidateQueries(["app-state"]),n&&n()},onError:async pe=>{i(!1),s(pe.message)}}):d(ie,{onSuccess:async pe=>{i(!1);const ge=Bf().user;if(ge&&(pe==null?void 0:pe.length)>1){const _e=pe[1];window.location.assign(`/user/${ge}/${_e}`)}},onError:async pe=>{i(!1),s(pe.message)}})},T=async({servername:k,user_options:x})=>(await Zo.post("/server",{servername:k,user_options:x})).data,m=async({servername:k,user_options:x})=>(await Zo.put(`/server/${k}`,{servername:k,user_options:x})).data,{mutate:d}=lc({mutationFn:T,retry:1}),{mutate:y}=lc({mutationFn:m,retry:1});return U.useEffect(()=>{u!=null&&u.name&&(u!=null&&u.user_options)&&(a(u.name),S({...u.user_options}))},[u==null?void 0:u.name,u==null?void 0:u.user_options,S]),U.useEffect(()=>{f&&s(f.message)},[f,s]),C.jsxs("form",{id:"app-form",onSubmit:h(v),className:"form",children:[C.jsxs(_r,{children:[C.jsx(Rr,{htmlFor:"display_name",required:!0,children:"Display Name"}),C.jsx(Er,{name:"display_name",control:c,rules:rp,render:({field:{ref:k,...x}})=>C.jsx($o,{...x,id:"display_name",autoFocus:!0})}),((_=p.display_name)==null?void 0:_.message)&&C.jsx(Wh,{errors:[p.display_name.message]})]}),C.jsxs(_r,{children:[C.jsx(Rr,{htmlFor:"description",children:"Description"}),C.jsx(Er,{name:"description",control:c,render:({field:{ref:k,...x}})=>C.jsx(rN,{...x,id:"description"})})]}),C.jsxs(_r,{children:[C.jsx(Rr,{htmlFor:"framework",required:!0,children:"Framework"}),C.jsx(Er,{name:"framework",control:c,rules:rp,render:({field:{ref:k,...x}})=>C.jsx(tN,{...x,id:"framework",options:[{value:"",label:"Select..."},{value:"panel",label:"Panel"},{value:"bokeh",label:"Bokeh"},{value:"streamlit",label:"Streamlit"},{value:"voila",label:"Voila"},{value:"plotlydash",label:"PlotlyDash"},{value:"gradio",label:"Gradio"},{value:"jupyterlab",label:"JupyterLab"},{value:"custom",label:"Custom Command"}]})}),((A=p.framework)==null?void 0:A.message)&&C.jsx(Wh,{errors:[p.framework.message]})]}),C.jsxs(_r,{children:[C.jsx(Rr,{htmlFor:"filepath",children:"Filepath"}),C.jsx(Er,{name:"filepath",control:c,render:({field:{ref:k,...x}})=>C.jsx($o,{...x,id:"filepath"})})]}),C.jsxs(_r,{children:[C.jsx(Rr,{htmlFor:"conda_env",children:"Conda Environment"}),C.jsx(Er,{name:"conda_env",control:c,render:({field:{ref:k,...x}})=>C.jsx($o,{...x,id:"conda_env"})})]}),C.jsxs(_r,{children:[C.jsx(Rr,{htmlFor:"custom_command",children:"Custom Command"}),C.jsx(Er,{name:"custom_command",control:c,render:({field:{ref:k,...x}})=>C.jsx($o,{...x,id:"custom_command"})})]}),C.jsxs(mg,{children:[C.jsx(io,{id:"cancel-btn",type:"button",variant:"secondary",onClick:t,children:"Cancel"}),C.jsx(io,{id:"submit-btn",type:"submit",disabled:o,children:"Submit"})]})]})},zA=({id:e,title:t,description:n,thumbnail:r,framework:o,url:i,ready:s=!1})=>{const l=ea(),[a,u]=U.useState(!1),[,f]=Hn(Oi),[c,h]=U.useState(!1),[S,p]=U.useState(!1),v=async({id:_})=>await Zo.delete(`/server/${_}`),{mutate:T}=lc({mutationFn:v,retry:1}),m=()=>{u(!0),T({id:e},{onSuccess:async()=>{u(!1),h(!1),l.invalidateQueries(["app-state"])},onError:async _=>{u(!1),f(_.message)}})},d=[{id:"start",title:"Start",onClick:()=>window.location.assign(i),disabled:s,visible:!1},{id:"stop",title:"Stop",onClick:()=>console.log(`Stopping app with id: ${e}`),disabled:!s,visible:!1},{id:"edit",title:"Edit",onClick:()=>p(!0),visible:!0},{id:"delete",title:"Delete",onClick:()=>h(!0),visible:!0}],y=C.jsxs(C.Fragment,{children:[C.jsxs("p",{className:"w-[400px] mb-6",children:["Are you sure you want to delete ",C.jsx("b",{children:t}),"? This action is permanent and cannot be reversed."]}),C.jsxs(mg,{children:[C.jsx(io,{id:"cancel-btn",variant:"secondary",onClick:()=>h(!1),children:"Cancel"}),C.jsx(io,{id:"delete-btn",variant:"primary",onClick:()=>m(),disabled:a,children:"Delete"})]})]});return C.jsxs("div",{className:"card",id:`card-${e}`,tabIndex:0,children:[C.jsxs("div",{className:"card-header-media",children:[C.jsxs("div",{className:"card-header-menu",children:[C.jsx(eN,{id:`card-menu-${e}`,items:d}),c&&C.jsx(ac,{title:`Delete ${t}`,setIsOpen:h,body:y}),S&&C.jsx(ac,{title:`Edit ${t}`,setIsOpen:p,body:C.jsx(Xg,{id:e,onCancel:()=>p(!1),onSubmit:()=>p(!1)})})]}),C.jsx("div",{className:"card-header-img",children:r?C.jsx("img",{src:r,alt:"App thumb"}):void 0})]}),C.jsx("div",{className:"card-header",children:C.jsx("h3",{className:"font-bold",children:C.jsx("a",{href:i,children:t})})}),C.jsx("div",{className:"card-body",children:C.jsx("p",{className:"text-sm",children:n})}),C.jsx("div",{className:"card-footer",children:C.jsx(nN,{id:`tag-${e}`,children:o})})]})},dp=({appType:e="My",filter:t})=>{const[n]=Hn(Df),[,r]=Hn(Oi),[o,i]=U.useState([]),{isLoading:s,error:l,data:a}=Ff({queryKey:["app-state"],queryFn:()=>Zo.get("/server/").then(u=>u.data).then(u=>u),enabled:!!n.user});return U.useEffect(()=>{if(!s&&a){const u=t.toLowerCase();i(()=>EA(a,e).filter(f=>{var c,h;return f.name.toLowerCase().includes(u)||((c=f.description)==null?void 0:c.toLowerCase().includes(u))||((h=f.framework)==null?void 0:h.toLowerCase().includes(u))}))}},[s,a,e,t]),U.useEffect(()=>{r(l?l.message:void 0)},[l,r]),C.jsxs(C.Fragment,{children:[C.jsxs("div",{className:"container grid grid-cols-12 flex flex-align-center pb-12",children:[C.jsx("div",{className:"col-span-1",children:C.jsxs("h2",{className:"whitespace-nowrap font-bold",children:[e," Apps"]})}),C.jsx("div",{className:"col-span-10",children:C.jsx("hr",{className:"spacer"})}),C.jsx("div",{className:"col-span-1 flex justify-end",children:C.jsxs("h2",{className:"whitespace-nowrap font-bold",children:[o.length," apps"]})})]}),C.jsx("div",{className:"container grid pb-12",children:o.length>0?C.jsx("div",{className:"flex flex flex-row flex-wrap gap-4",children:o.map(u=>C.jsx(zA,{id:u.id,title:u.name,description:u.description,thumbnail:u.thumbnail,framework:u.framework,url:u.url,ready:u.ready},`app-${u.id}`))}):C.jsx("div",{children:"No apps available"})})]})},HA=Ug.create({baseURL:"/api",headers:{"Content-Type":"application/json",Authorization:"Authorization: token 13748105b2e64b9082e10f802b04f523"},params:{_xsrf:Bf().xsrf_token}}),WA=()=>{const[e]=Hn(Df),[,t]=Hn(Oi),[n,r]=U.useState([]),{isLoading:o,error:i,data:s}=Ff({queryKey:["service-data"],queryFn:()=>HA.get("/services").then(a=>a.data).then(a=>a),enabled:!!e.user}),l=(a,u)=>{u?window.open(a,"_blank"):window.location.assign(a)};return U.useEffect(()=>{!o&&s&&r(()=>RA(s,e.user))},[o,s,e.user]),U.useEffect(()=>{t(i?i.message:void 0)},[i,t]),C.jsxs(C.Fragment,{children:[C.jsxs("div",{className:"container grid grid-cols-12 flex flex-align-center pb-12",children:[C.jsx("div",{className:"col-span-1",children:C.jsx("h2",{className:"whitespace-nowrap font-bold",children:"Services"})}),C.jsx("div",{className:"col-span-10",children:C.jsx("hr",{className:"spacer"})}),C.jsx("div",{className:"col-span-1 flex justify-end",children:C.jsxs("h2",{className:"whitespace-nowrap font-bold",children:[n.length," services"]})})]}),C.jsx("div",{className:"container grid pb-12",children:n.length>0?C.jsx("div",{className:"flex flex flex-row flex-wrap gap-4",children:n.map((a,u)=>C.jsx(io,{id:`service-${u}`,variant:"secondary",style:{minWidth:"180px"},onClick:()=>{l(a.url,a.external)},children:a.name},`service-${u}`))}):C.jsx("div",{children:"No services available"})})]})},hp=()=>{const[e,t]=U.useState(!1),[n]=Hn(Oi),[r,o]=U.useState(""),i=s=>{const l=s.target;o(l.value)};return C.jsxs(C.Fragment,{children:[C.jsxs("div",{className:"container grid grid-cols-12 pb-12",children:[C.jsx("div",{className:"md:col-span-2 xs:col-span-12",children:C.jsx("h1",{className:"text-3xl font-bold",children:"Home"})}),C.jsx("div",{className:"md:col-span-8 xs:col-span-8",children:C.jsx($o,{id:"search",placeholder:"Search...","aria-label":"Search for an app",className:"w-full mt-0",onChange:i})}),C.jsxs("div",{className:"md:col-span-2 xs:col-span-4 flex justify-end",children:[C.jsx(io,{id:"create-app",onClick:()=>{t(!0)},children:"Create App"}),e&&C.jsx(ac,{title:"Create New App",setIsOpen:t,body:C.jsx(Xg,{onCancel:()=>t(!1)})})]})]}),n&&C.jsx("div",{className:"container grid grid-cols-12 pb-2",children:C.jsx("div",{className:"col-span-12",children:C.jsx(qk,{id:"alert-notification",type:"error",children:n})})}),C.jsx(WA,{}),C.jsx(dp,{appType:"My",filter:r}),C.jsx(dp,{appType:"Shared",filter:r})]})},QA=new vk,qA=()=>{const[,e]=Hn(Df);return U.useEffect(()=>{e(Bf())},[e]),C.jsx(Ok,{client:QA,children:C.jsx("div",{children:C.jsx("main",{className:"my-6",children:C.jsxs(u_,{children:[C.jsx(Qu,{path:"/home",element:C.jsx(hp,{})}),C.jsx(Qu,{path:"/",element:C.jsx(hp,{})})]})})})})};Ja.createRoot(document.getElementById("root")).render(C.jsx(ne.StrictMode,{children:C.jsx(f_,{basename:"/hub",children:C.jsx(nk,{children:C.jsx(qA,{})})})}));
