+++ ui/src/components/index.ts
@@ -1,6 +1,4 @@
 export { default as Alert } from './alert/alert';
-export { default as AppCard } from './app-card/app-card';
-export { default as AppForm } from './app-form/app-form';
 export { default as ButtonGroup } from './button-group/button-group';
 export { default as Button } from './button/button';
 export { default as ContextMenu } from './context-menu/context-menu';
